% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{aggregatets}
\alias{aggregatets}
\title{Aggregate a time series}
\usage{
aggregatets(ts, on = "minutes", k = 1, tz = "GMT")
}
\arguments{
\item{ts}{xts object to aggregate.}

\item{on}{character, indicating the time scale in which "k" is expressed. Possible values are: "secs", "seconds", "mins", "minutes", "hours", "days", "weeks".}

\item{k}{positive integer, indicating the number of periods to aggregate over. E.g. to aggregate a 
xts object to the 5 minute frequency set k=5 and on="minutes".}

\item{tz}{time zone used}
}
\value{
An xts object containing the aggregated time series.
}
\description{
Function returns aggregated time series as xts object. 
It can handle irregularly spaced timeseries and returns a regularly spaced one.
Use univariate timeseries as input for this function, and check out \code{\link{aggregateTrades}}
and \code{\link{aggregateQuotes}} to aggregate Trade or Quote data objects.
}
\details{
The timestamps of the new time series are the closing times and/or days of the intervals. 
E.g. for a weekly aggregation the new timestamp is the last day in that particular week (namely sunday).

In case of previous tick aggregation, 
for on = "seconds"/"minutes"/"hours",
the element of the returned series with e.g. timestamp 09:35:00 contains 
the last observation up to that point, excluding the value at 09:35:00 itself.

Please Note:
In case an interval is empty these NA's are filled by the function \code{na.locf}
(last observation carried forward) from the zoo package.
}
\examples{
#load sample price data
ts <- sample_tdata$PRICE

#Previous tick aggregation to the 5-minute sampling frequency:
tsagg5min <- aggregatets(ts, on = "minutes", k = 5)
head(tsagg5min)
#Previous tick aggregation to the 30-second sampling frequency:
tsagg30sec <- aggregatets(ts, on = "seconds", k = 30)
tail(tsagg30sec)

}
\author{
Jonathan Cornelissen and Kris Boudt
}
\keyword{data}
\keyword{manipulation}
