% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{autoCorrelationOfNoise}
\alias{autoCorrelationOfNoise}
\title{Autocorrelation of noise estimation}
\usage{
autoCorrelationOfNoise(pData, kn, lags)
}
\description{
Autocorrelation of noise estimation
}
