% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_data_handling.R
\name{BFMalgorithm}
\alias{BFMalgorithm}
\title{This function is not exported, but the documentation is here to help users}
\usage{
BFMalgorithm(tData, qData, backwardsWindow, forwardsWindow, plot, tz)
}
\description{
This function is not exported, but the documentation is here to help users
}
\keyword{internal}
