% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalRealizedMeasures.R
\name{cholCovrMRCov}
\alias{cholCovrMRCov}
\title{#' @keywords internal
zgamma <- function (x, y, gamma_power) {
  if (x^2 < y) {
    out <- abs(x)^gamma_power
  } else {
    if (gamma_power == 1) {
      out <- 1.094 * sqrt(y)
    }
    if (gamma_power == 2) {
      out <- 1.207 * y
    }
    if (gamma_power == 4/3) {
      out <- 1.129 * y^(2/3)
    }
  }
  
  return(out)
}}
\usage{
cholCovrMRCov(returns, delta = 0.1, theta = 1)
}
\description{
#' @keywords internal
zgamma <- function (x, y, gamma_power) {
  if (x^2 < y) {
    out <- abs(x)^gamma_power
  } else {
    if (gamma_power == 1) {
      out <- 1.094 * sqrt(y)
    }
    if (gamma_power == 2) {
      out <- 1.207 * y
    }
    if (gamma_power == 4/3) {
      out <- 1.129 * y^(2/3)
    }
  }
  
  return(out)
}
}
\keyword{internal}
