% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driftBursts.R
\name{driftBursts}
\alias{driftBursts}
\title{driftBursts
  Drift Bursts}
\usage{
driftBursts(
  pData,
  testTimes = seq(34260, 57600, 60),
  preAverage = 5,
  ACLag = -1L,
  meanBandwidth = 300L,
  varianceBandwidth = 900L,
  parallelize = FALSE,
  nCores = NA,
  warnings = TRUE
)
}
\arguments{
\item{testTimes}{A \code{numeric} containing the times at which to calculate the tests. The standard of \code{seq(34260, 57600, 60)} 
denotes calculating the test-statistic once per minute, i.e. 390 times for a typical 6.5 hour trading day from 9:31:00 to 16:00:00. See details. Default is seq(34260, 57600, 60.}

\item{preAverage}{An \code{integer} denoting the length of pre-averaging window for the log-prices. Default is 5}

\item{ACLag}{A positive \code{integer} greater than 1 denoting how many lags are to be used for the HAC estimator of the variance - the default
of \code{-1} denotes using an automatic lag selection algorithm for each iteration. Default is -1L}

\item{meanBandwidth}{An \code{integer} denoting the bandwidth for the left-sided exponential kernel for the mean. Default is 300L}

\item{varianceBandwidth}{An \code{integer} denoting the bandwidth for the left-sided exponential kernel for the variance. Default is 900L}

\item{parallelize}{A \code{logical} to determine whether to parallelize the underlying C++ code (Using OpenMP). Default is FALSE}

\item{nCores}{An \code{integer} denoting the number of cores to use for calculating the code when paralellized. 
If this argument is not provided, sequential evaluation will be used even though \code{parallelize} is TRUE. Default is NA}

\item{warnings}{A \code{logical} denoting whether warnings should be shown. Default is TRUE}

\item{timestamps}{Either: A \code{numeric} of \code{timestamps} for the trades in seconds after midnight.
Or: \code{NULL}, when the time argument is \code{NULL}, the \code{logprices} argument must be an \code{xts} object.}

\item{logPrices}{A \code{numeric} or \code{xts} object containing the log-prices.}

\item{sessionStart}{A \code{double} to denote the start of the trading session in seconds after midnight. Default is 34200}

\item{sessionEnd}{A \code{double} to denote the end of the trading session in seconds after midnight. Default is 57600}
}
\value{
An object of class \code{DBH} and \code{list} containing the series of the drift burst hypothesis test-statistic as well as the estimated spot drift and variance series. 
The list also contains some information such as the variance and mean bandwidths along with the pre-averaging setting and the amount of observations. 
Additionally, the list will contain information on whether testing happened for all \code{testTimes} entries.
}
\description{
Calculates the Test-Statistic for the Drift Burst Hypothesis
}
\details{
If the \code{testTimes} vector contains instructions to test before the first trade, or more than 15 minutes after the last trade, these entries will be deleted, as not doing so may cause crashes.
The test statistic is unstable before \code{max(meanBandwidth , varianceBandwidth)} seconds has passed.
If \code{timestamps} is provided and \code{logPrices} is an \code{xts} object, the indices of logPrices will be used regardless.
Note that using an \code{xts} logPrices argument is slower than using a \code{numeric} due to the creation of the timestamps from the index of the input. 
When using \code{xts} objects, be careful to use the correct time zones. For example, if I as a dane use the \code{"America/New_York"} time zone for my \code{xts} objects, I have to add 14400 to my testing times.
Same correction will have to be made to the \code{startTime} and \code{endTime} arguments in the plotting methods.
The lags from the Newey-West algorithm is increased by \code{2 * (preAveage-1)} due to the pre-averaging we know at least this many lags should be corrected for.
The maximum of 20 lags is also increased by this factor for the same reason.
}
\author{
Emil Sjoerup
}
