% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driftBursts.R
\name{getCriticalValues}
\alias{getCriticalValues}
\title{Get critical value for the drift burst hypothesis t-statistic}
\usage{
getCriticalValues(x, alpha = 0.95)
}
\arguments{
\item{x}{object of class \code{DBH}}

\item{alpha}{numeric denoting the confidence level for the critical value

The critical value is that of the test described in appendix B in Christensen Oomen Reno}
}
\description{
Method for DBH objects to calculate the critical value for the presence of a burst of drift.
}
