% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquidity_measures.R
\name{getLiquidityMeasures}
\alias{getLiquidityMeasures}
\title{Compute Liquidity Measure}
\usage{
getLiquidityMeasures(tqdata, win = 300, type = NULL)
}
\arguments{
\item{tqdata}{A \code{data.table} or xts object as in the \pkg{highfrequency} merged
trades and quotes data (that is included).}

\item{win}{A windows length for the forward-prices used for \sQuote{realized}
spread}

\item{type}{Legacy option. Default is to return all liquidity measures.}
}
\value{
A modified (enlarged) \code{data.table} with the new measures.
}
\description{
Liquidity Measures are in the \pkg{highfrequency} by Boudt,
Cornelissen and Payseur are computed, but this time in the context
of a \code{data.table} object which is much more efficient.
}
\details{
Note that this assumes a regular time grid. The Lee + Ready measure
uses two lags for the Tick Rule.

NOTE: xts or data.table should only contain one day of observations
}
\examples{
tqdata <- matchTradesQuotes(sample_tdata, sample_qdata)
res <- getLiquidityMeasures(tqdata)
head(res)
}
