% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_data_handling.R
\name{hasQty}
\alias{hasQty}
\title{Check for Trade, Bid, and Ask/Offer (BBO/TBBO), Quantity, and Price data}
\usage{
hasQty(x, which = FALSE)
}
\arguments{
\item{x}{data object}

\item{which}{disply position of match}
}
\description{
A set of functions to check for appropriate TBBO/BBO and price column
names within a data object, as well as the availability and
position of those columns.
}
