% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heavy_model_new.R
\name{HEAVYmodelNew}
\alias{HEAVYmodelNew}
\title{keywords internal
calcRecVarEq <- function(par, rm) {
  h <- rep(1, times = length(rm))
  h[1] <- mean(rm)
  for (ii in c(2:length(rm))) {
    h[ii] <- par[1] + par[2] * rm[ii-1] + par[3] * h[ii-1]
  }
  h
}
HEAVY Model estimation}
\usage{
HEAVYmodelNew(ret, rm, startingValues = NULL)
}
\arguments{
\item{ret}{(Log)-returns}

\item{rm}{realized measure of intraday stock price variation}

\item{startingValues}{alternative starting values}
}
\description{
This function calculates the High frEquency bAsed VolatilitY (HEAVY) model proposed in Shephard and Sheppard (2010).
}
\references{
Shephard, N. and K. Sheppard (2010). Realising the future: Forecasting with high frequency based volatility (HEAVY) models. Journal of Applied Econometrics 25, 197--231.
}
