% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{hfsim.do}
\alias{hfsim.do}
\title{simulate hf prices}
\usage{
hfsim.do(hfSimSpec)
}
\arguments{
\item{hfSimSpec}{object of type hfSimSpec created by \code{\link{createHFSimSpec}}}
}
\value{
a List containing the simulated data and additional information, e.g. underlying variance process, bursts in the drift or volatility, and if applicable noise term.
}
\description{
simulate hf prices
}
\author{
Emil Sjoerup
}
