% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{hfSim}
\alias{hfSim}
\title{simulate hf prices}
\usage{
hfSim(spec)
}
\arguments{
\item{spec}{object of type spec created by \code{\link{hfSimSpec}}}
}
\value{
an object of class \code{hfSim} which is a list containing the simulated data and any applicable additional information, e.g. underlying variance process, bursts in the drift or volatility.
}
\description{
simulate hf prices
}
\details{
The class \code{hfSim} has the following methods implemented:
\enumerate{
\item{\code{print}} prints the simulated prices and information about the spec and simulation
}
}
\author{
Emil Sjoerup
}
