% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfSimSpec.R
\name{hfSimSpec}
\alias{hfSimSpec}
\title{create high frequency simulation spec}
\usage{
hfSimSpec(volatilityModel = list(modelType = "constant", sigma = 0.2),
          driftModel = list(modelType = "none", drift = 0),
          jumpModel = list(modelType = "none"),
          diurnalModel = list(modelType = "none"),
          burstModel = list(driftModel = list(modelType = "none"), 
                            volModel = list(modelType = "none")),
          noiseModel = list(noiseType = "none"),
          timeSettings = list(tradingStart = 34200, tradingEnd = 57600, 
                              origin = "1970-01-01", sampling = "equidistant"),
          nSeries = 1,
          nDays = 1,
          nObs = 23401,
          discretize = FALSE,
          returnType = "DT")
}
\arguments{
\item{volatilityModel}{List containing the volatility model specification. Possible inputs are: 
\enumerate{
   \item \code{modelType} A string denoting which model type to use, available models and descriptions thereof are available through the \code{\link{listAvailableVolatilityModels}} function
   \item \code{sigma} A symmetric matrix denoting the variance-covariance matrix. Not used when \code{modelType} is "huangTauchen"
   \item \code{volOfVol} numeric of length 1 or \code{nSeries} denoting the volatility of the volatility process for one-factor models
   \item \code{meanReversion} numeric of or length 1 or \code{nSeries} denoting the strength of mean reversion for the volatility process in one-factor models
   \item \code{rho} numeric of length 1 or \code{nSeries} denoting the correlation between \eqn{w_{t}} and \eqn{b_{t}} in the Heston model. It is analogous for \eqn{w_{1}} and \eqn{w_{2}} in the Li Linton model
   \item \code{alpha1} numeric of length 1 or \code{nSeries} denoting the strength of mean reversion in the persistent volatility component in the Huang & Tauchen model
   \item \code{alpha2} numeric of length 1 or \code{nSeries} denoting the strength of mean reversion in the transient volatility component in the Huang & Tauchen model
   \item \code{beta0} numeric of length 1 or \code{nSeries} denoting the level of the total volatility component in the Huang & Tauchen model
   \item \code{beta1} numeric of length 1 or \code{nSeries} denoting the level of the persistent volatility component in the Huang & Tauchen model
   \item \code{beta2} numeric of length 1 or \code{nSeries} denoting the level of the transient volatility component in the Huang & Tauchen model
   \item \code{phi} numeric of length 1 or \code{nSeries} denoting the strength of the feedback in the transient volatility factor in the Huang & Tauchen model
   \item \code{rho1} numeric of length 1 or \code{nSeries} denoting the leverage correlations between \eqn{w_{1,t}} and \eqn{w_{2,t}} in the Huang & Tauchen model
   \item \code{volJumps} numeric of length 1 or \code{nSeries} denoting the intensity of the jumps in the volatility factor in the Li & Linton model
   \item \code{lambda} numeric of length 1 or \code{nSeries} denoting the intensity of the poisson process that governs jumps in the price and volatility in the Li & Linton model
   \item \code{mu1} numeric of length 1 or \code{nSeries} denoting the long term mean of the price in the Li & Linton model
   \item \code{kappa1} numeric of length 1 or \code{nSeries} denoting the mean reversion of the price process in the Li & Linton model
}}

\item{driftModel}{List containing the drift model specification. Possible inputs are:
 \enumerate{
   \item \code{modelType} A string denoting which model type to use, available models and descriptions thereof are available through the \code{\link{listAvailableDriftModels}} function
   \item \code{drift} numeric of length 1 or \code{nSeries} denoting the daily drift in case modelType = "constant", and the long term mean of drift in the Vasicek model. Default is 0.
   \item \code{meanReversion} numeric of length 1 or \code{nSeries} denoting the mean reversion of drift. Default is 2.
   \item \code{driftVol} numeric of length 1 or \code{nSeries} denoting the volatility of drift in case modelType = "vasicek".
}}

\item{jumpModel}{List containing the jump model specification. Possible inputs are:
\enumerate{
   \item \code{modelType} A string denoting which model type to use, available models and descriptions thereof are available through the \code{\link{listAvailableJumpModels}} function
   \item \code{jumpComponent} numeric of length 1 or \code{nSeries} between 0 and 1 denoting how much of the total sigma of the model that should come from the jump variation.
}}

\item{diurnalModel}{List containing the diurnal model specification. Possible inputs are:
\enumerate{
   \item \code{modelType} A string denoting which model type to use, available models and descriptions thereof are available through the \code{\link{listAvailableDiurnalModels}} function.
   \item \code{C} numeric of length 1 denoting the level of the diurnal pattern. (Usually calibrated to make sure that when integrating over the diurnal factor, it gives 1)
   \item \code{A} numeric of length 1 denoting the A value of the diurnal pattern. This is used to control the level of the diurnal factor during the beginning of the day. Default is 0.75
   \item \code{B} numeric of length 1 denoting the B value of the diurnal pattern. This is used to control the level of the diurnal factor during the end of the day. Default is 0.25
   \item \code{a} numeric of length 1 denoting the a value of the diurnal pattern. This is used to control the steepness of the drop during the beginning of the day. Default is 10
   \item \code{b} numeric of length 1 denoting the b value of the diurnal pattern. This is used to control the steepness of the increase during the end of the day. Default is 10
}}

\item{burstModel}{List containing the burst model specifications. Possible inputs are: 
\enumerate{
   \item driftModel list with entries 
   \enumerate{
         \item \code{modelType} A string denoting which model type to use, available models and descriptions thereof are available through the \code{\link{listAvailableBurstModels}} function
         \item \code{alpha} numeric of length one denoting the alpha exponent in the drift burst. Default is 0.65 
         \item \code{a} numeric of length one denoting the a scalar in the drift burst this can be used to control the severity of the drift burst. Default is 1 
         \item \code{burstInterval} numeric of length two denoting the interval when the drift burst is to take place. Default is \code{c(15/32, 17/32)} 
   }
   \item volModel list with entries 
   \enumerate{
         \item \code{modelType} A string denoting which model type to use, available models and descriptions thereof are available through the \code{\link{listAvailableBurstModels}} function
         \item \code{beta} (Used only when modelType is "singularityBurst") A numeric of length one denoting the beta exponent in the volatility burst. Default is 0.4 
         \item \code{b} (Used only when modelType is "singularityBurst") A numeric of length one denoting the b scalar in the volatility burst this can be used to control the severity of the volatility burst. Default is 1 
         \item \code{burstMultiplier} (Used only when modelType is "constantBurst") A numeric of length one denoting the severity of the volatility burst. Default is 3 denoting a temporary three-fold increase of volatility
         \item \code{burstInterval} A numeric of length two when the volatility burst is to take place. Default is \code{c(15/32, 17/32)} 
   }
}}

\item{noiseModel}{List containing the noise model specification. Possible inputs are: 
\enumerate{
   \item \code{modelType} A string denoting which model type to use, available models and descriptions thereof are available through the \code{\link{listAvailableNoiseModels}} function.
   \item \code{signalToNoise} A numeric of either length 1, nrow(sigma) or nSeries denoting the signal to noise ratio of the noise.
   \item \code{variance} A numeric of either length 1, nrow(sigma) or nSeries denoting the variance of the noise when the variance is constant.
}}

\item{timeSettings}{List containing the settings for the time in the simulation. Possible inputs are:
\enumerate{
   \item \code{tradingStart} A numeric of length one denoting in seconds after midnight the start of trading. Default is 34200 which corresponds to 09:30
   \item \code{tradingEnd} A numeric of length one denoting in seconds after midnight the start of trading. Default is 57600 which corresponds to 16:00
   \item \code{origin} A character which can be coerced to a date, denoting the first day of the simulated series. Default is "1970-01-01"
   \item \code{sampling} A character denoting which sampling scheme to use. Currently only "equidistant" is implemented.
}}

\item{nSeries}{Integer-valued numeric of length one denoting how many series to simulate. This is overwritten in case nrow of the variance covariance matrix is greater than one. Then, nSeries will be set to match \code{nrow(sigma)}}

\item{nDays}{Integer-valued numeric of length one denoting how many days to simulate the data over.}

\item{nObs}{Integer-valued numeric of length one denoting how many days to simulate the data over.}

\item{discretize}{Logical denoting whether to discretize the prices. The prices are discretized using log(round(100 * exp(prices)) / 100). i.e. to nearest cent.}

\item{returnType}{character determining the type of object to use for the storage of returns, prices, jumps etc. Defaults to "DT", which corresponds to \code{data.table} any other input will return \code{xts} objects.}
}
\value{
an object of type "highfrequencySimSpec"
}
\description{
create high frequency simulation spec
}
\author{
Emil Sjoerup
}
