\name{highfrequency-package}
\alias{highfrequency-package}
\alias{highfrequency}
\docType{package}
\title{
highfrequency: Toolkit for the analysis of highfrequency financial data in R.
}
\description{
The highfrequency package contains an extensive toolkit for the use of highfrequency financial data in R. It contains functionality to manage, clean and match highfrequency trades and quotes data. Furthermore, it enables users to: calculate easily various liquidity measures, estimate and forecast volatility, and investigate microstructure noise and intraday periodicity.}
\details{
\tabular{ll}{
Package: \tab highfrequency\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-06-01\cr
License: \tab \cr
}
}

\author{
Jonathan Cornelissen, Kris Boudt, Scott Payseur

Maintainer: Jonathan Cornelissen <Jonathan.cornelissen@econ.kuleuven.be>
}

\keyword{ package }

\examples{
# see users manual
}