\name{highfrequencyGSOC-package}
\alias{highfrequencyGSOC-package}
\alias{highfrequencyGSOC}
\docType{package}
\title{
Additional functionality for the highfrequency package
}
\description{
Additional functionality for the highfrequency package: jump test, standard error
}
\details{
\tabular{ll}{
Package: \tab highfrequencyGSOC\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-07-02\cr
License: \tab GPL\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Giang Nguyen, Kris Boudt, Jonathan Cornelissen

Maintainer: Giang Nguyen <hagiang3108@gmail.com>
~~ The author and/or maintainer of the package ~~
}
\references{

}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
}
\examples{

}
