% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{huangTauchen}
\alias{huangTauchen}
\title{Huang and Tauchen 2005 model
(we don't include drift here) (sigma_ut is calculated elsewhere too!)
sigma_ut * nu_t (rho_1 * dWt_1 + rho_2 * dWt_2 + sqrt(1 - rho_1 ^2 - rho^2) * dWt_3)
nu_t^2 = s-exp(beta_0 + beta_1 * nu_t_1^2 + beta_2 * nu_t_2^2)
dnu_1_t^2 = alpha_1  * nu_1_t^2 * dt + dWt_1
dnu_2_t^2 = alpha_2 * nu_2_t^2 * dt + (1 + phi nu_2_t^2) * dW2_t
sigma_ut = C + A*exp(-a*t) + B*exp(-b*(1-t))}
\usage{
huangTauchen(model, nObs, nSeries, nDays, dt)
}
\description{
Huang and Tauchen 2005 model
(we don't include drift here) (sigma_ut is calculated elsewhere too!)
sigma_ut * nu_t (rho_1 * dWt_1 + rho_2 * dWt_2 + sqrt(1 - rho_1 ^2 - rho^2) * dWt_3)
nu_t^2 = s-exp(beta_0 + beta_1 * nu_t_1^2 + beta_2 * nu_t_2^2)
dnu_1_t^2 = alpha_1  * nu_1_t^2 * dt + dWt_1
dnu_2_t^2 = alpha_2 * nu_2_t^2 * dt + (1 + phi nu_2_t^2) * dW2_t
sigma_ut = C + A*exp(-a*t) + B*exp(-b*(1-t))
}
\keyword{internal}
