% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realizedMeasures.R
\name{ICov}
\alias{ICov}
\title{Estimators of the integrated covariance}
\description{
This documentation page functions as a point of reference to quickly look up the estimators of the integrated covariance provided in the \pkg{highfrequency} package.

The implemented estimators are:

Realized covariance \code{\link{rCov}}

Realized bipower covariance \code{\link{rBPCov}}

Hayashi-Yoshida realized covariance \code{\link{rHYCov}}

Realized kernel covariance \code{\link{rKernelCov}}

Realized outlyingness-weighted covariance \code{\link{rOWCov}}

Realized threshold covariance \code{\link{rThresholdCov}}

Realized two-scale covariance \code{\link{rTSCov}}

Robust realized two-scale covariance \code{\link{rRTSCov}}

Subsampled realized covariance \code{\link{rAVGCov}}

Realized semi-covariance \code{\link{rSemiCov}}

Modulated Realized covariance \code{\link{rMRCov}}

Realized Cholesky covariance \code{\link{rCholCov}}

Beta-adjusted realized covariance \code{\link{rBACov}}
}
\seealso{
\code{\link{IVar}} for a list of implemented estimators of the integrated variance.
}
