% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realizedMeasures.R
\name{IVar}
\alias{IVar}
\title{Estimators of the integrated variance}
\description{
This documentation page functions as a point of reference to quickly look up the estimators of the integrated variance provided in the \pkg{highfrequency} package.

The implemented estimators are:
Realized Variance \code{\link{rRVar}}

Median realized variance \code{\link{rMedRVar}}

Minimum realized variance \code{\link{rMinRVar}}

Realized quadpower variance \code{\link{rQPVar}}

Realized multipower variance \code{\link{rMPVar}}

Realized semivariance \code{\link{rSVar}} 

Note that almost all estimators in the list in \code{\link{ICov}} also work yield estimates of the integrated variance on the diagonals.
}
\seealso{
\code{\link{ICov}} for a list of implemented estimators of the integrated covariance.
}
