% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{knChooseREMEDI}
\alias{knChooseREMEDI}
\title{REMEDI tuning parameter
function to choose the tuning parameter, kn in REMEDI estimation}
\usage{
knChooseREMEDI(
  pData,
  correctTime = FALSE,
  jumpsIndex = NULL,
  knMax = 10,
  tol = 0.05,
  size = 3,
  lower = 2,
  upper = 5
)
}
\description{
REMEDI tuning parameter
function to choose the tuning parameter, kn in REMEDI estimation
}
