% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfsimSpec.R
\name{listAvailableBurstModels}
\alias{listAvailableBurstModels}
\title{List the available burst models for simulations}
\usage{
listAvailableBurstModels()
}
\value{
This function returns the available burst models in a matrix
}
\description{
List the available burst models for simulations
}
\author{
Emil Sjoerup
}
