% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfSimSpec.R
\name{listAvailableDiurnalModels}
\alias{listAvailableDiurnalModels}
\title{List the available diurnal models for simulations}
\usage{
listAvailableDiurnalModels()
}
\value{
This function returns the available diurnal models in a matrix
}
\description{
List the available diurnal models for simulations
}
\author{
Emil Sjoerup
}
