% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfsimSpec.R
\name{listAvailableDriftModels}
\alias{listAvailableDriftModels}
\title{List the available drift models for simulations}
\usage{
listAvailableDriftModels()
}
\value{
This function returns the available drift models in a matrix
}
\description{
List the available drift models for simulations
}
\author{
Emil Sjoerup
}
