% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{listAvailableJumpModels}
\alias{listAvailableJumpModels}
\title{List the available jump models for simulations}
\usage{
listAvailableJumpModels()
}
\value{
This function returns the available Jump models in a matrix
}
\description{
List the available jump models for simulations
}
