% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfsimSpec.R
\name{listAvailableNoiseModels}
\alias{listAvailableNoiseModels}
\title{List the available noise models for simulations}
\usage{
listAvailableNoiseModels()
}
\value{
This function returns the available noise models in a matrix
}
\description{
List the available noise models for simulations
}
\author{
Emil Sjoerup
}
