% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfSimSpec.R
\name{listAvailableVolatilityModels}
\alias{listAvailableVolatilityModels}
\title{List the available volatility models for simulations}
\usage{
listAvailableVolatilityModels()
}
\value{
This function returns the available volatility models in a matrix
}
\description{
List the available volatility models for simulations
}
\author{
Emil Sjoerup
}
