% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realizedMeasures.R
\name{listCholCovEstimators}
\alias{listCholCovEstimators}
\title{Utility function listing the available estimators for the CholCov estimation}
\usage{
listCholCovEstimators()
}
\value{
This function returns a character vector containing the available estimators.
}
\description{
Utility function listing the available estimators for the CholCov estimation
}
