% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lltc}
\alias{lltc}
\title{LLTC Data}
\format{xts object}
\usage{
lltc
}
\description{
Tick data for LLTC 2011/07/01, cleaned with \code{tradesCleanup}.
}
\examples{
data(lltc)
plot(lltc)

}
\keyword{datasets}
