% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandling.R
\name{makeRMFormat}
\alias{makeRMFormat}
\title{DEPRECATED 
use \code{\link{spreadPrices}}}
\usage{
makeRMFormat(data)
}
\arguments{
\item{data}{DEPRECATED}
}
\description{
DEPRECATED 
use \code{\link{spreadPrices}}
}
