% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{minRQ}
\alias{minRQ}
\title{An estimator of integrated quarticity from applying the minimum operator on blocks of two returns.}
\usage{
minRQ(rData, alignBy = NULL, alignPeriod = NULL, makeReturns = FALSE)
}
\arguments{
\item{rData}{an xts or data.table object containing returns or prices, possibly for multiple assets over multiple days}

\item{alignBy}{a string, align the tick data to "seconds"|"minutes"|"hours"}

\item{alignPeriod}{an integer, align the tick data to this many [seconds|minutes|hours].}

\item{makeReturns}{boolean, should be TRUE when rData contains prices instead of returns. FALSE by  default.}
}
\value{
numeric
}
\description{
Function returns the minRQ, defined in Andersen et al. (2012).

Assume there is \eqn{N} equispaced returns in period \eqn{t}. Let \eqn{r_{t,i}} be a return (with \eqn{i=1, \ldots,N}) in period \eqn{t}.

Then, the minRQ is given by
\deqn{
  \mbox{minRQ}_{t}=\frac{\pi N}{3 \pi - 8} \left(\frac{N}{N-1}\right) \sum_{i=1}^{N-1} \mbox{min}(|r_{t,i}| ,|r_{t,i+1}|)^4
}
}
\examples{
rq <- minRQ(rData = sampleTDataMicroseconds[, list(DT, PRICE)], alignBy = "minutes", 
            alignPeriod = 5, makeReturns = TRUE)
rq
}
\references{
Andersen, T. G., D. Dobrev, and E. Schaumburg (2012). Jump-robust volatility estimation using nearest neighbor truncation. Journal of Econometrics, 169(1), 75- 93.
}
\author{
Giang Nguyen, Jonathan Cornelissen and Kris Boudt
}
