% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{minRV}
\alias{minRV}
\title{minRV}
\usage{
minRV(rData, alignBy = NULL, alignPeriod = NULL, makeReturns = FALSE)
}
\arguments{
\item{rData}{an xts or data.table object containing returns or prices, possibly for multiple assets over multiple days}

\item{alignBy}{a string, align the tick data to "seconds"|"minutes"|"hours".}

\item{alignPeriod}{an integer, align the tick data to this many [seconds|minutes|hours].}

\item{makeReturns}{boolean, should be TRUE when rData contains prices instead of returns. FALSE by default.}
}
\value{
In case the input is an xts object with data from one day, a numeric of same length as the number of assets. 
If the input data spans multiple days and is in xts format, an xts will be returned. 
If the input data is a data.table object, the function returns a data.table with the same column names as the input data, containing the date and the realized measures
}
\description{
Function returns the minRV, defined in Andersen et al. (2009).

Let \eqn{r_{t,i}} be a return (with \eqn{i=1,\ldots,M}) in period \eqn{t}.

Then, the minRV is given by
\deqn{
\mbox{minRV}_{t}=\frac{\pi}{\pi - 2}\left(\frac{M}{M-1}\right) \sum_{i=1}^{M-1} \mbox{min}(|r_{t,i}| ,|r_{t,i+1}|)^2
}
}
\examples{
minrv <- minRV(rData = sampleTDataMicroseconds[, list(DT, PRICE)], alignBy = "minutes",
               alignPeriod = 5, makeReturns = TRUE)
minrv 

}
\references{
Andersen, T. G., D. Dobrev, and E. Schaumburg (2012). Jump-robust volatility estimation using nearest neighbor truncation. Journal of Econometrics, 169 (1), 75-93.
}
\author{
Jonathan Cornelissen and Kris Boudt
}
\keyword{volatility}
