% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_jump_tests.R
\name{mukp}
\alias{mukp}
\title{## mukp: to use when p,k different from range [4,6]}
\usage{
mukp(p, k, t = 1e+06)
}
\description{
## mukp: to use when p,k different from range [4,6]
}
\keyword{internal}
