% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{noZeroPrices}
\alias{noZeroPrices}
\title{Delete the observations where the price is zero}
\usage{
noZeroPrices(tData)
}
\arguments{
\item{tData}{an xts or data.table object at least containing a column "PRICE".}
}
\value{
an xts or data.table object depending on input
}
\description{
Function deletes the observations where the price is zero.
}
\author{
Jonathan Cornelissen and Kris Boudt
}
\keyword{cleaning}
