% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{noZeroQuotes}
\alias{noZeroQuotes}
\title{Delete the observations where the bid or ask is zero}
\usage{
noZeroQuotes(qData)
}
\arguments{
\item{qData}{an xts or data.table object at least containing the columns "BID" and "OFR".}
}
\value{
xts object or data.table depending on type of input
}
\description{
Function deletes the observations where the bid or ask is zero.
}
\author{
Jonathan Cornelissen and Kris Boudt
}
\keyword{cleaning}
