% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_model.R
\name{plot.HARmodel}
\alias{plot.HARmodel}
\title{Plotting method for HARmodel objects}
\usage{
\method{plot}{HARmodel}(x, ...)
}
\arguments{
\item{x}{an object of class \code{HARmodel}}

\item{...}{extra arguments, see details}
}
\description{
Plotting method for HARmodel objects
}
\details{
The plotting method has the following optional parameter:
\itemize{
\item{\code{legend.loc}}{ A string denoting the location of the legend passed on to \code{addLegend} of the \pkg{xts} package}
}
}
