% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEAVYmodel.R
\name{plot.HEAVYmodel}
\alias{plot.HEAVYmodel}
\title{Plotting method for HEAVYmodel objects}
\usage{
\method{plot}{HEAVYmodel}(x, ...)
}
\arguments{
\item{x}{an object of class \code{HEAVYmodel}.}

\item{...}{extra arguments, see details.}
}
\description{
Plotting method for HEAVYmodel objects
}
\details{
The plotting method has the following optional parameter:
\itemize{
\item{\code{legend.loc}}{ A string denoting the location of the legend passed on to \code{addLegend} of the \pkg{xts} package}
\item{\code{type}}{ A string denoting the type of lot to be made. If \code{type} is \code{"condVar"} the fitted values of the conditional variance of the returns
is shown. If \code{type} is different from \code{"condVar"}, the fitted values of the realized measure is shown. Default is \code{"condVar"}}
}
}
