% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_model.R
\name{predict.HARmodel}
\alias{predict.HARmodel}
\title{Predict method for objects of type \code{HARmodel}}
\usage{
\method{predict}{HARmodel}(object, newdata = NULL, warnings = TRUE, backtransform = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{HARmodel}}

\item{newdata}{new data to use for forecasting}

\item{warnings}{logical denoting whether to display warnings}

\item{backtransfrom}{if the model is estimated with transformation this parameter can be set to transform the prediction back into variance
The possible values are \code{"simple"} which means inverse of transformation, i.e. \code{exp} when log-transformation is applied. If using log transformation,
the option \code{"parametric"} can also be used to transform back. The parametric method adds a correction}
}
\description{
Predict method for objects of type \code{HARmodel}
}
