% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HARmodel.R
\name{print.HARmodel}
\alias{print.HARmodel}
\title{Printing method for \code{HARmodel} objects}
\usage{
\method{print}{HARmodel}(x, ...)
}
\arguments{
\item{x}{object of type \code{HARmodel}}

\item{...}{extra options}
}
\description{
Printing method for \code{HARmodel} objects
}
\details{
The printing method has the extra option \code{digits} which can be used to set the number of digits for printing pass \code{lag} to determine the maximum order of the Newey West estimator. Default is \code{22}
}
