% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{rBPCov}
\alias{rBPCov}
\title{Realized BiPower Covariance}
\usage{
rBPCov(
  rData,
  cor = FALSE,
  alignBy = NULL,
  alignPeriod = NULL,
  makeReturns = FALSE,
  makePsd = FALSE
)
}
\arguments{
\item{rData}{an xts or data.table object containing returns or prices, possibly for multiple assets over multiple days}

\item{cor}{boolean, in case it is TRUE, the correlation is returned. FALSE by default.}

\item{alignBy}{a string, align the tick data to "seconds"|"minutes"|"hours".}

\item{alignPeriod}{an integer, align the tick data to this many [seconds|minutes|hours].}

\item{makeReturns}{boolean, should be TRUE when rData contains prices instead of returns. FALSE by default.}

\item{makePsd}{boolean, in case it is TRUE, the positive definite version of rBPCov is returned. FALSE by default.}
}
\value{
in case the input is and contains data from one day, an N by N matrix is returned. If the data is a univariate xts object with multiple days, an xts is returned. 
If the data is multivariate and contains multiple days (xts or data.table), the function returns a list containing N by N matrices. Each item in the list has a name which corresponds to the date for the matrix.
}
\description{
Function returns the Realized BiPower Covariance (rBPCov), 
 defined in Barndorff-Nielsen and Shephard (2004).
 
 Let \eqn{r_{t,i}} be an intraday \eqn{N x 1} return vector and \eqn{i=1,...,M} 
 the number of intraday returns.
 
 The rBPCov is defined as the process whose value at time \eqn{t}
 is the \eqn{N}-dimensional square matrix with \eqn{k,q}-th element equal to
 \deqn{
   \mbox{rBPCov}[k,q]_t = \frac{\pi}{8} \bigg( \sum_{i=2}^{M} 
                                              \left|
                                                r_{(k)t,i} + r_{(q)t,i} \right| \ \left| r_{(k)t,i-1} + r_{(q)t,i-1} \right|   \\
                                              - \left| r_{(k)t,i}  - r_{(q)t,i} \right| \ \left|
                                                r_{(k)t,i-1} - r_{(q)t,i-1} \right|  \bigg),
 }
 where \eqn{r_{(k)t,i}} is the
 \eqn{k}-th component of the return vector \eqn{r_{i,t}}.
}
\examples{
# Realized Bipower Variance/Covariance for a price series aligned   
# at 5 minutes.
 
# Univariate: 
rbpv <- rBPCov(rData = sampleTData[, list(DT, PRICE)], alignBy ="minutes", 
               alignPeriod = 5, makeReturns = TRUE) 
# Multivariate: 
rbpc <- rBPCov(rData = sampleOneMinuteData, makeReturns = TRUE, makePsd = TRUE)
rbpc
 
}
\references{
Barndorff-Nielsen, O. and N. Shephard (2004). Measuring the impact of
jumps in multivariate price processes using bipower covariation. Discussion
paper, Nuffield College, Oxford University.
}
\author{
Jonathan Cornelissen, Kris Boudt, and Emil Sjoerup
}
\keyword{volatility}
