% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalRealizedMeasures.R
\name{RBPCov_bi}
\alias{RBPCov_bi}
\title{# Check data:
#' @keywords internal
rdatacheck <- function (rData, multi = FALSE) {
  if ((dim(rData)[2] < 2) & (multi)) {
    stop("Your rData object should have at least 2 columns")
  }
}}
\usage{
RBPCov_bi(ts1, ts2)
}
\description{
# Check data:
#' @keywords internal
rdatacheck <- function (rData, multi = FALSE) {
  if ((dim(rData)[2] < 2) & (multi)) {
    stop("Your rData object should have at least 2 columns")
  }
}
}
\keyword{internal}
