% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{REMEDI}
\alias{REMEDI}
\title{REMEDI
estimates auto-covariance of market-microstructure noise}
\usage{
REMEDI(
  pData,
  kn = 1,
  lags = 1,
  correctTime = FALSE,
  jumpsIndex = NULL,
  makeCorrelation = FALSE
)
}
\arguments{
\item{pData}{xts or data.table containing the log-prices of the asset.}

\item{kn}{numeric of length 1 determining the tuning parameter kn}

\item{lags}{numeric containing integer values indicating}

\item{correctTime}{logical indicating whether to use the time-adjusted REMEDI measure, default is FALSE}

\item{jumpsIndex}{Index of jumps}

\item{makeCorrelation}{logical indicating whether to transform the autocovariances into autocorrelations.}
}
\description{
REMEDI
estimates auto-covariance of market-microstructure noise
}
\references{
remedi paper, muzafer's paper
}
\keyword{autocorrelation}
\keyword{autocovariance}
\keyword{microstructure}
\keyword{noise}
