% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{ReMeDIAsymptoticVariance}
\alias{ReMeDIAsymptoticVariance}
\title{Asymptotic variance of ReMeDI estimator}
\usage{
ReMeDIAsymptoticVariance(pData, kn, lags, phi, i)
}
\arguments{
\item{pData}{\code{xts} or \code{data.table} containing the log-prices of the asset}

\item{kn}{numeric of length 1 determining the tuning parameter kn this controls the lengths of the non-overlapping interval in the ReMeDI estimation}

\item{lags}{numeric containing integer values indicating the lags for which to estimate the (co)variance}

\item{phi}{tuning parameter phi}

\item{i}{tuning parameter i}
}
\value{
a list with compontents ReMeDI and asympVar
}
\description{
Asymptotic variance of ReMeDI estimator
}
