\name{rKernel.available}
\alias{rKernel.available}

\title{Available Kernels}
\description{
Returns a vector of the available kernels.
}
\usage{
rKernel.available()
}

\references{
Ole E. Barndorff-Nielsen, Peter Reinhard Hansen, Asger Lunde, and Neil Shephard. Regular and modified kernel-based estimators of integrated variance: The case with independent noise. \emph{Working Paper}, 2004.
 }

\author{ Scott Payseur <scott.payseur@gmail.com>}

\seealso{ \code{\link{rKernelCov}} }
\examples{
rKernel.available()
}
\keyword{volatility}
