% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realizedMeasures.R
\name{rMedRQ}
\alias{rMedRQ}
\title{DEPRECATED}
\usage{
rMedRQ(rData, alignBy = NULL, alignPeriod = NULL, makeReturns = FALSE)
}
\arguments{
\item{rData}{DEPRECATED}

\item{alignBy}{DEPRECATED}

\item{alignPeriod}{DEPRECATED}

\item{makeReturns}{DEPRECATED}
}
\description{
DEPRECATED
}
