% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandling.R
\name{rmNegativeSpread}
\alias{rmNegativeSpread}
\title{Delete entries for which the spread is negative}
\usage{
rmNegativeSpread(qData)
}
\arguments{
\item{qData}{an \code{xts} object at least containing the columns "BID" and "OFR".}
}
\value{
\code{data.table} or \code{xts} object
}
\description{
Function deletes entries for which the spread is negative.
}
\examples{
rmNegativeSpread(sampleQDataRaw)

}
\author{
Jonathan Cornelissen, Kris Boudt and Onno Kleen
}
\keyword{cleaning}
