% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realizedMeasures.R
\name{rMRC}
\alias{rMRC}
\title{DEPRECATED rMRC}
\usage{
rMRC(pData, pairwise = FALSE, makePsd = FALSE, theta = 0.8, ...)
}
\arguments{
\item{pData}{DEPRECATED}

\item{pairwise}{DEPRECATED}

\item{makePsd}{DEPRECATED}

\item{theta}{DEPRECATED}

\item{...}{DEPRECATED}
}
\description{
DEPRECATED USE \code{\link{rMRCov}}
}
