% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{rmTradeOutliersUsingQuotes}
\alias{rmTradeOutliersUsingQuotes}
\title{Delete transactions with unlikely transaction prices}
\usage{
rmTradeOutliersUsingQuotes(tData, qData)
}
\arguments{
\item{tData}{a data.table or xts object containing the time series data, with at least the column "PRICE", containing the transaction price (ONE DAY ONLY).}

\item{qData}{a data.table or xts object containing the time series data with at least the columns "BID" and "OFR", containing the bid and ask prices (ONE DAY ONLY).}
}
\value{
xts or data.table object depending on input
}
\description{
Function deletes entries with prices that are above the ask plus the bid-ask spread.
Similar for entries with prices below the bid minus the bid-ask spread.
}
\details{
Note: in order to work correctly, the input data of this function should be
cleaned trade (tData) and quote (qData) data respectively.
}
\author{
Jonathan Cornelissen, Kris Boudt and Onno Kleen
}
\keyword{cleaning}
