% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realizedMeasures.R
\name{rSV}
\alias{rSV}
\title{DEPRECATED}
\usage{
rSV(rData, alignBy = NULL, alignPeriod = NULL, makeReturns = FALSE)
}
\arguments{
\item{rData}{DEPRECATED USE \code{\link{rSVar}}}

\item{alignBy}{DEPRECATED USE \code{\link{rSVar}}}

\item{alignPeriod}{DEPRECATED USE \code{\link{rSVar}}}

\item{makeReturns}{DEPRECATED USE \code{\link{rSVar}}}
}
\description{
DEPRECATED USE \code{\link{rSVar}}
}
