% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{RTQ}
\alias{RTQ}
\title{Calculate the realized tripower quarticity}
\usage{
RTQ(rData)
}
\arguments{
\item{rData}{a zoo/xts object containing all returns in period t for one asset.}
}
\value{
numeric
}
\description{
Calculate the realized tripower quarticity
}
