% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realizedMeasures.R
\name{RV}
\alias{RV}
\title{DEPRECATED
DEPRECATED USE \code{\link{rRVar}}}
\usage{
RV(rData)
}
\arguments{
\item{rData}{DEPRECATED USE \code{\link{rRVar}}}
}
\description{
DEPRECATED
DEPRECATED USE \code{\link{rRVar}}
}
