% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_qdata_microseconds}
\alias{sample_qdata_microseconds}
\title{Sample of cleaned quotes for stock XXX for 2 days measured in microseconds}
\format{
data.table object
}
\usage{
sample_qdata_microseconds
}
\description{
A data.table object containing the raw quotes for the imaginary stock XXX for 2 days. This is the cleaned version of the data sample \code{\link{sample_qdataraw_microseconds}}, using \code{quotesCleanup}.
}
\keyword{datasets}
