% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_qdataraw}
\alias{sample_qdataraw}
\title{Sample of raw quotes for stock XXX for 1 day}
\format{xts object}
\usage{
sample_qdataraw
}
\description{
An imaginary xts object containing the raw quotes for stock XXX for 1 day, in the typical NYSE TAQ database format.
This is the cleaned version of the data sample \code{\link{sample_tdataraw}}, using \code{tradesCleanup}.
}
\keyword{datasets}
