% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleQDataRaw_microseconds}
\alias{sampleQDataRaw_microseconds}
\title{Sample of raw quotes for stock XXX for 2 days measured in microseconds}
\format{
data.table object
}
\usage{
sampleQDataRaw_microseconds
}
\description{
An imaginary xts object containing the raw quotes for stock XXX for 2 days, in the typical NYSE TAQ database format.
}
\keyword{datasets}
