% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleReturns5Min}
\alias{sampleReturns5Min}
\title{Sample returns data}
\format{
A large \code{xts} object.
}
\usage{
sampleReturns5Min
}
\description{
EUR/USD returns from January to September 2004
}
\keyword{datasets}
