% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleTData}
\alias{sampleTData}
\title{Sample of cleaned trades for stock XXX for 1 day}
\format{
A large \code{xts} object.
}
\usage{
sampleTData
}
\description{
An xts object containing the trades for the imaginary stock XXX for 1 day, in the typical NYSE TAQ database format.
This is the cleaned version of the data sample \code{\link{sampleTDataRaw}}, using \code{tradesCleanup}.
}
\keyword{datasets}
