% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampletDataRaw}
\alias{sampletDataRaw}
\title{Sample of raw trades for stock XXX for 1 day}
\format{
A large \code{xts} object.
}
\usage{
sampletDataRaw
}
\description{
An imaginary xts object containing the raw trades for stock XXX for 1 day, in the typical NYSE TAQ database format.
}
\keyword{datasets}
