% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleMultiTradeData}
\alias{sampleMultiTradeData}
\title{Multivariate tick by tick data}
\format{
A \code{data.table} object
}
\usage{
sampleMultiTradeData
}
\description{
Cleaned Tick by tick data for a sector ETF, called \code{ETF} and two stock components of that ETF,
these stocks are named \code{AAA} and \code{BBB}.
}
\keyword{datasets}
