% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleOneMinuteData}
\alias{sampleOneMinuteData}
\title{One minute data}
\format{
A \code{data.table} object
}
\usage{
sampleOneMinuteData
}
\description{
One minute data price of one stock and a market proxy. This is data from the US market.
}
\keyword{datasets}
