% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleQDataMicroseconds}
\alias{sampleQDataMicroseconds}
\title{Sample of cleaned quotes for stock XXX for 2 days measured in microseconds}
\format{
data.table object
}
\usage{
sampleQDataMicroseconds
}
\description{
A data.table object containing the raw quotes for the imaginary stock XXX for 2 days. This is the cleaned version of the data sample \code{\link{sampleQDataRawMicroseconds}}, using \code{quotesCleanup}.
}
\examples{
\dontrun{
# The code to create the sampleQDataMicroseconds dataset from raw data is
sampleQDataMicroseconds <- quotesCleanup(qDataRaw = sampleQDataRawMicroseconds,
                                         exchanges = "N", type = "standard", report = FALSE)
}
}
\keyword{datasets}
