% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleQDataRawMicroseconds}
\alias{sampleQDataRawMicroseconds}
\title{Sample of raw quotes for stock XXX for 2 days measured in microseconds}
\format{
data.table object
}
\usage{
sampleQDataRawMicroseconds
}
\description{
An data.table object containing the raw quotes for stock XXX for 2 days, in the typical NYSE TAQ database format.
}
\keyword{datasets}
