% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleReal5MinPrices}
\alias{sampleReal5MinPrices}
\title{Sample of imaginary price data for 61 days}
\format{
xts object
}
\usage{
sampleReal5MinPrices
}
\description{
An xts object containing the 5-min aggregated imaginary price series for the trading
days between 2005-03-04 and 2005-06-01.
}
\keyword{datasets}
