% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleTDataEurope}
\alias{sampleTDataEurope}
\title{European data}
\format{
A \code{data.table} object
}
\usage{
sampleTDataEurope
}
\description{
Trade data of one stock on one day in the European stock market.
}
\keyword{datasets}
