% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleTDataRaw}
\alias{sampleTDataRaw}
\title{Sample of raw trades for stock XXX for 2 days}
\format{
A data.table object.
}
\usage{
sampleTDataRaw
}
\description{
An imaginary \code{data.table} object containing the raw trades the pseudonymized stock XXX for 2 days, in the typical NYSE TAQ database format.
}
\keyword{datasets}
