% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbux}
\alias{sbux}
\title{Starbucks Data}
\format{xts object}
\usage{
sbux
}
\description{
Tick data for Starbucks 2011/07/01, cleaned with \code{tradesCleanup}.
}
\examples{
data(sbux)
plot(sbux) 

}
\keyword{datasets}
