% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SP500RM}
\alias{SP500RM}
\title{SP500 Realized Measures calculated with 5 minute sampling}
\format{A large \code{xts} object.}
\source{
\url{http://public.econ.duke.edu/~ap172/code.html}
}
\usage{
SP500RM
}
\description{
Realized measures from the SP500 index from April 1997 to August 2013.
}
\references{
Bollerslev, T., A. J. Patton, and R. Quaedvlieg, 2016, Exploiting the Errors: A Simple Approach for Improved Volatility Forecasting, Journal of Econometrics, 192, 1-18.
}
\keyword{datasets}
