% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPYRM}
\alias{SPYRM}
\title{SPY realized measures}
\format{
A \code{data.table} object
}
\usage{
SPYRM
}
\description{
Realized measures for the SPY ETF calculated at 1 and 5 minute sampling.
}
\note{
The CLOSE column is NOT the official close price, but simply the last recorded price of the day. Thus, this may be slightly different from other sources.
}
\keyword{datasets}
