% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HARmodel.R
\name{summary.HARmodel}
\alias{summary.HARmodel}
\title{Summary for \code{HARmodel} objects}
\usage{
\method{summary}{HARmodel}(object, ...)
}
\arguments{
\item{object}{An object of class \code{HARmodel}}

\item{...}{pass \code{lag} to determine the maximum order of the Newey West estimator. Default is \code{22}}
}
\value{
A modified \code{summary.lm}
}
\description{
Summary for \code{HARmodel} objects
}
