% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{autoSelectExchangeQuotes}
\alias{autoSelectExchangeQuotes}
\title{Retain only data from the stock exchange with the highest volume}
\usage{
autoSelectExchangeQuotes(qData, printExchange = TRUE)
}
\arguments{
\item{qData}{a \code{data.table} or \code{xts} object with at least a column "EX", indicating the exchange symbol 
and columns "BIDSIZ" and "OFRSIZ", indicating 
the volume available at the bid and ask respectively.}

\item{printExchange}{indicates whether the chosen exchange is printed on the console, default is TRUE.
The possible exchanges are:
\itemize{
\item A: AMEX
\item N: NYSE
\item B: Boston
\item P: Arca
\item C: NSX
\item T/Q: NASDAQ
\item D: NASD ADF and TRF
\item X: Philadelphia
\item I: ISE
\item M: Chicago
\item W: CBOE
\item Z: BATS
}}
}
\value{
\code{data.table} or \code{xts} object depending on input
}
\description{
Function returns an \code{xts} object containing only observations 
of the exchange with highest
value for the sum of "BIDSIZ" and "OFRSIZ", i.e. the highest quote volume.
}
\examples{
autoSelectExchangeQuotes(sampleQDataRaw)

}
\author{
Jonathan Cornelissen, Kris Boudt, Onno Kleen, and Emil Sjoerup
}
\keyword{cleaning}
