% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{autoSelectExchangeTrades}
\alias{autoSelectExchangeTrades}
\title{Retain only data from the stock exchange with the highest trading volume}
\usage{
autoSelectExchangeTrades(tData)
}
\arguments{
\item{tData}{an xts object with at least a column "EX", 
indicating the exchange symbol and "SIZE", 
indicating the trade volume. The chosen exchange is printed on the console.
\itemize{
\item A: AMEX
\item N: NYSE
\item B: Boston
\item P: Arca
\item C: NSX
\item T/Q: NASDAQ
\item D: NASD ADF and TRF
\item X: Philadelphia
\item I: ISE
\item M: Chicago
\item W: CBOE
\item Z: BATS
}}
}
\value{
data.table or xts object depending on input
}
\description{
Function returns a data.table or xts object containing only observations of the 
exchange with the highest value for the variable "SIZE", 
i.e. the highest trade volume.
}
\examples{
autoSelectExchangeTrades(sampletDataRawMicroseconds)

}
\author{
Jonathan Cornelissen, Kris Boudt and Onno Kleen
}
\keyword{cleaning}
