% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpTests.R
\name{BNSjumpTest}
\alias{BNSjumpTest}
\title{Barndorff-Nielsen and Shephard (2006) tests for the presence of jumps in the price series.}
\usage{
BNSjumpTest(
  rData,
  IVestimator = "BV",
  IQestimator = "TP",
  type = "linear",
  logTransform = FALSE,
  max = FALSE,
  alignBy = NULL,
  alignPeriod = NULL,
  makeReturns = FALSE,
  alpha = 0.975
)
}
\arguments{
\item{rData}{either an \code{xts} or a \code{data.table} containing the log-returns or prices of a single asset, possibly over multiple days-}

\item{IVestimator}{can be chosen among jump robust integrated variance estimators: BV, rMinRV, rMedRV and corrected threshold bipower variation (CTBV). If CTBV is chosen, an argument of \eqn{startV}, start point of auxiliary estimators in threshold estimation (Corsi et al. (2010) can be included. BV by default.}

\item{IQestimator}{can be chosen among jump robust integrated quarticity estimators: TP, QP, rMinRQ and rMedRQ. TP by default.}

\item{type}{a method of BNS testing: can be linear or ratio. Linear by default.}

\item{logTransform}{boolean, should be TRUE when QVestimator and IVestimator are in logarithm form. FALSE by default.}

\item{max}{boolean, should be TRUE when max adjustment in SE. FALSE by default.}

\item{alignBy}{character, indicating the time scale in which \code{alignPeriod} is expressed. Possible values are: "secs", "seconds", "mins", "minutes","hours".
To aggregate based on a 5 minute frequency, set \code{alignPeriod} to 5 and \code{alignBy} to "minutes".}

\item{alignPeriod}{positive numeric, indicating the number of periods to aggregate over. E.g. to aggregate
based on a 5 minute frequency, set \code{alignPeriod} to 5 and \code{alignBy} to "minutes".}

\item{makeReturns}{boolean, should be TRUE when pData contains prices. FALSE by default.}

\item{alpha}{numeric of length one with the significance level to use for the jump test(s). Defaults to 0.975.}
}
\value{
list or \code{xts}in case the input prices span more than one day.
}
\description{
This test examines the presence of jumps in highfrequency price series. It is based on theory of Barndorff- Nielsen and Shephard (BNS). The null hypothesis is no jumps. 
Depending on the choice of estimator (integrated variance (IVestimator), integrated quarticity (IQestimator)), mechanism (linear, ratio) and adjustment (logarithm), the function returns the result.
Function returns three outcomes: 1.z-test value 2.critical value(with confidence level of 95\%) and 3.pvalue of the test. 
Assume there is \eqn{N} equispaced returns in period \eqn{t}. 

Assume the Realized variance (RV), IVestimator and IQestimator are based on \eqn{N} equispaced returns. 

Let \eqn{r_{t,i}} be a return (with \eqn{i=1, \ldots,N}) in period \eqn{t}. 

Then the BNSjumpTest is given by: 
\deqn{
\mbox{BNSjumpTest}= \frac{RV - IVestimator}{\sqrt{(\theta-2)\frac{1}{N} {IQestimator}}}
}
in which, \eqn{IVestimator} can be: bipower variance (BV), rMinRV, rMedRV. 
\eqn{IQestimator} can be: tripower quarticity (TP), quadpower quarticity (QP), rMinRQ, rMedRQ.

\eqn{\theta}: depends on IVestimator (Huang and Tauchen (2005)).
}
\details{
The theoretical framework underlying jump test is that the logarithmic price process \eqn{X_t} belongs to the class of Brownian semimartingales, which can be written as:
\deqn{
\mbox{X}_{t}=  \int_{0}^{t} a_udu + \int_{0}^{t}\sigma_{u}dW_{u} + Z_t
}
where \eqn{a} is the drift term, \eqn{\sigma} denotes the spot volatility process, \eqn{W} is a standard Brownian motion and \eqn{Z} is a jump process defined by:
\deqn{
\mbox{Z}_{t}=  \sum_{j=1}^{N_t}k_j
}
where \eqn{k_j} are nonzero random variables. The counting process can be either finite or infinite for finite or infinite activity jumps.

Since the realized volatility converges to the sum of integrated variance and jump variation, while the robust IVestimator converges to the integrated variance, 
it follows that the difference between  \eqn{RV_{t,N}} and the IVestimator captures the jump part only, and this observation underlines the BNS test for jumps (Theodosiou and Zikes, 2009).
}
\examples{
bns <- BNSjumpTest(sampleTData[, list(DT, PRICE)], IVestimator= "rMinRV",
                   IQestimator = "rMedRQ", type= "linear", makeReturns = TRUE)
bns

}
\references{
Barndorff-Nielsen, O. E., and Shephard, N. (2006). Econometrics of testing for jumps in financial economics using bipower variation. \emph{Journal of Financial Econometrics}, 4, 1-30. 

Corsi, F., Pirino, D., and Reno, R. (2010). Threshold bipower variation and the impact of jumps on volatility forecasting. \emph{Journal of Econometrics}, 159, 276-288.

Huang, X., and Tauchen, G. (2005). The relative contribution of jumps to total price variance. \emph{Journal of Financial Econometrics}, 3, 456-499.

Theodosiou, M., and Zikes, F. (2009). A comprehensive comparison of alternative tests for jumps in asset prices. Unpublished manuscript, Graduate School of Business, Imperial College London.
}
\author{
Giang Nguyen, Jonathan Cornelissen, Kris Boudt, and Emil Sjoerup
}
\keyword{BNSjumpTest}
\keyword{highfrequency}
