% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{createHFSimSpec}
\alias{createHFSimSpec}
\title{create high frequency simulation spec}
\usage{
createHFSimSpec(
  volatilityModel = list(modelType = "constant", variance = 0.2, burstMultiplier = 3,
    includeDiurnality = FALSE, diurnalModel = list(C = 0.88929198, A = 0.75, B = 0.25, a
    = 10, b = 10)),
  driftModel = list(modelType = "constant", drift = 0),
  jumpModel = list(modelType = "none", jumpComponent = 1/5, jumpTime = c(0.5, 17/32)),
  noiseModel = list(noiseType = "constant with interday change", signalToNoiseRatio =
    2),
  nSeries = 1,
  nDays = 1,
  nObs = 23401,
  discretize = FALSE,
  timeSettings = list(tradingStart = 34200, tradingEnd = 57600, origin = "1970-01-01",
    sampling = "equidistant")
)
}
\arguments{
\item{jumpModel}{Object of type list with entries \enumerate{
   \item modelType 
   \item jumpComponent double in (0,1) denoting how much of the total variance of the model that should come from the jump variation.
   \item
   \item
   \item
   \item
}}

\item{model}{Object of type list with the entries: \enumerate{
   \item modelType A string denoting which model type to use, available models are available through the \code{listAvailableVolatilityModels} function
   \item driftModel A strig denoting which (if any) drift model type to use, available models are available through the \code{listAvailableDriftModels} function
   \item drift
   \item volatility the total volatility (standard deviation) of the simulated model if using jumps, the jump component will lower this argument
}}
}
\description{
create high frequency simulation spec
}
