% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{exchangeHoursOnly}
\alias{exchangeHoursOnly}
\title{Extract data from an xts object for the Exchange Hours Only}
\usage{
exchangeHoursOnly(
  data,
  marketOpen = "09:30:00",
  marketClose = "16:00:00",
  tz = NULL
)
}
\arguments{
\item{data}{a data.table or xts object containing the time series data. 
Multiple days of input are allowed.}

\item{dayBegin}{character in the format of \"HH:MM:SS\",
specifying the starting hour, minute and second of an exhange
trading day.}

\item{dayEnd}{character in the format of \"HH:MM:SS\^",
specifying the closing hour, minute and second of an exchange
trading day.}
}
\value{
xts or data.table object depending on input
}
\description{
The function returns data within exchange trading hours
"dayBegin" and "dayEnd". By default, dayBegin and dayEnd
are set to "09:30:00" and "16:00:00" respectively (see Brownlees and Gallo (2006) for more information on good choices for these arguments).
}
\examples{
exchangeHoursOnly(sampleTDataRawMicroseconds)
}
\references{
Brownlees, C.T. and Gallo, G.M. (2006). Financial econometric analysis at ultra-high frequency: Data handling concerns. Computational Statistics & Data Analysis, 51, pages 2232-2245.
}
\author{
Jonathan Cornelissen, Kris Boudt and Onno Kleen.
}
\keyword{cleaning}
