% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driftBursts.R
\name{getCriticalValues}
\alias{getCriticalValues}
\alias{getCriticalValues.DBH}
\title{Get critical value for the drift burst hypothesis t-statistic}
\usage{
getCriticalValues(x, alpha = 0.95)
}
\arguments{
\item{x}{object of class \code{DBH}}

\item{alpha}{numeric denoting the confidence level for the critical value. Possible values are \code{c(0.9 0.95 0.99 0.995 0.999 0.9999)}}
}
\description{
Method for DBH objects to calculate the critical value for the presence of a burst of drift.
The critical value is that of the test described in appendix B in Christensen Oomen Reno
}
\references{
Christensen, K., Oomen, R., and Reno, R. (2018) The Drift Burst Hypothesis (Working paper).
}
\author{
Emil Sjoerup
}
