% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heavy_model.R
\name{HEAVYmodel}
\alias{HEAVYmodel}
\title{HEAVY Model estimation}
\usage{
HEAVYmodel(ret, rm, startingValues = NULL)
}
\arguments{
\item{ret}{a vector of demeaned returns}

\item{rm}{a vector of realized stock market variation}

\item{startingValues}{a vector of alternative starting values: first three arguments for variance equation and last three arguments for measurement equation.}
}
\value{
The function outputs an object of class \code{HEAVYmodel}, a list containing
\itemize{
  \item coefficients = estimated coefficients
  \item se = robust standard errors based on inverted Hessian matrix
  \item residuals = the residuals in the return equation
  \item llh = the two-component log-likelihood values
  \item varCondVariances = conditional variances in the variance equation
  \item RMCondVariances = conditional variances in the RM equation
  \item data = the input data
}
}
\description{
This function calculates the High frEquency bAsed VolatilitY (HEAVY) model proposed in Shephard and Sheppard (2010).
}
\examples{

# Calculate annualized returns in percentages
logReturns <- 100 * makeReturns(SPYRM$CLOSE)[-1]

# Returns are assumed to be demeaned
logReturns <- logReturns - mean(logReturns)

# Due to return calculation, the first observation is missing
estimatedHEAVY <- HEAVYmodel(logReturns, SPYRM$RK5[-1] * 10000)

# Examine tThe estimated coefficients and robust standard errors
estimatedHEAVY

# Calculate iterative multi-step-ahead forecasts
predict(estimatedHEAVY, stepsAhead = 12)
}
\references{
Shephard, N. and K. Sheppard (2010). Realising the future: Forecasting with high frequency based volatility (HEAVY) models. Journal of Applied Econometrics 25, 197--231.
}
\author{
Onno Kleen
}
