% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highfrequency.R
\docType{package}
\name{highfrequency-package}
\alias{highfrequency}
\alias{highfrequency-package}
\title{highfrequency: Tools for Highfrequency Data Analysis}
\description{
The \pkg{highfrequency} package provides numerous tools for analyzing high frequency financial data. It has five main goals:

\itemize{ 
\item Provide functionality to manage, clean and match highfrequency trades and quotes data, 
\item calculate various liquidity measures, 
\item estimate and forecast volatility, 
\item detect price jumps and 
\item investigate microstructure noise and intraday periodicity.
}

To learn more about \pkg{highfrequency}, start with the vignettes:
`browseVignettes(package = "highfrequency")`
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonathancornelissen/highfrequency}
}

}
\author{
Kris Boudt, Jonathan Cornelissen, Scott Payseur

Maintainer: Kris Boudt <Kris.Boudt@ugent.be>

Contributors: Giang Nguyen, Onno Kleen, Emil Sjoerup

Thanks: We would like to thank Brian Peterson, Chris Blakely, Eric Zivot and Maarten Schermer. We are also grateful to Dirk Eddelbuettel for his extraordinary support as a mentor during the Google Summmer of Code 2019.
}
