\name{highfrequency-package}
\alias{highfrequency-package}
\alias{highfrequency}
\docType{package}
\title{
Tools For Highfrequency Data Analysis
}
\description{
Provide functionality to manage, clean and match highfrequency
    trades and quotes data, calculate various liquidity measures, estimate and
    forecast volatility, and investigate microstructure noise and intraday
    periodicity.}
\details{
\tabular{ll}{
Package: \tab highfrequency\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2014-08-31\cr
License: \tab GPL (>= 2) \cr
}
}

\author{
Jonathan Cornelissen,  Kris Boudt, Scott Payseur

Maintainer: Kris Boudt <Kris.Boudt@econ.kuleuven.be>

Contributors: Giang Nguyen, Maarten Schermers.

Thanks: A special thanks for additional contributions from Chris Blakely, Brian Peterson, Eric Zivot and GSoC.
}

\keyword{ package }

\examples{
# see users manual
}
