% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highfrequency.R
\docType{package}
\name{highfrequency-package}
\alias{highfrequency}
\alias{highfrequency-package}
\title{highfrequency: Tools for Highfrequency Data Analysis}
\description{
The \pkg{highfrequency} package provides numerous tools for analyzing high-frequency financial data, including functionality to:
\itemize{
\item Clean, handle, and manage high frequency trades and quotes data.
\item Calculate liquidity measures
\item Calculate (multivariate) realized measures of the distribution of high-frequency returns
\item Estimate models for realized measures of volatility and the corresponding forecasts
\item Detect jumps in prices
\item Analyze market microstructure noise in asset prices
\item Estimate spot volatility and drift as well as analyze intraday periodicity of spot volatility
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonathancornelissen/highfrequency}
}

}
\author{
Kris Boudt, Jonathan Cornelissen, Onno Kleen, Scott Payseur, Emil Sjoerup
Maintainer: Kris Boudt <Kris.Boudt@ugent.be>

Contributors: Giang Nguyen

Thanks: We would like to thank Brian Peterson, Chris Blakely, Dirk Eddelbuettel, Maarten Schermer, and Eric Zivot
}
