\name{highfrequencyGSOC-package}
\alias{highfrequencyGSOC-package}
\alias{highfrequencyGSOC}
\docType{package}
\title{
Additional functionality for the highfrequency package
}

\description{Extra functionality for the highfrequency package, written by Giang, during the Google Summer of Code.}


\author{
Giang Nguyen, Kris Boudt, Jonathan Cornelissen
Maintainer: Giang Nguyen <hagiang3108@gmail.com>
}