% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realizedMeasures.R
\docType{data}
\name{ICov}
\alias{ICov}
\title{Estimators of the integrated covariance}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
ICov
}
\description{
This documentation page functions as a point of reference to quickly look up the estimators of the integrated covariance provided in the \pkg{highfrequency} package.

The implemented estimators are:

Realized Covariance \code{\link{rCov}}

Subsampled Realized Covariance \code{\link{rAVGCov}}

Modulated Realized Covariance \code{\link{rMRCov}}
}
\keyword{datasets}
