% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump_tests.R
\name{intradayJumpTest}
\alias{intradayJumpTest}
\title{General framework for testing for jumps on an intraday basis}
\usage{
intradayJumpTest(
  pData,
  volEstimator = "RM",
  driftEstimator = "none",
  alpha = 0.95,
  ...,
  on = "minutes",
  k = 5,
  marketOpen = "09:30:00",
  marketClose = "16:00:00",
  tz = "GMT"
)
}
\arguments{
\item{pData}{(currently only xts) of the price data in levels. This data can (and should in some cases) be tick-level data. This data can span more than one day}

\item{alpha}{numeric significance level to use for the jump tests.}

\item{...}{used internally. Don't set this parameter.}

\item{testType}{type of test to use. Currently supports "LM" test of Lee and Mykland (2018), and "FoF" as "Fact or Friction" test from Christensen, Oomen, and Podolskij (2014)}

\item{testingTimes}{a character vector of times to test for jumps, for example  \code{c("11:30", "11:35", "14:30")}. 
This argument can also be a numeric vector containing times to test in seconds after midnight. For example, using \code{seq(39200, 57600, 300)}, 
where testing will take place every five minutes, starting fifty minutes after opening. (On when open is 09:30). This argument is not used in the FoF type test.}

\item{windowSize}{the size of the window to test for jumps in, in minutes. Default = 5. Not used in the FoF type test.}

\item{K}{In the LM test, K, is the amount of windowSizes to use for estimation of the local variance. 
In case of the FoF test K will be used as M in eqn: 21 in the paper, which changes the meaning into how many non-overlapping pre-averaged returns should be used in each test.}

\item{theta}{numeric parameter in determining the pre-averaging horizon. Default = 0.5. The pre-averaging horizon is \code{round(theta * sqrt(n))}, where n is the number of observations. 
This parameter will also help determine the testing times as the test is done on non-overlapping pre-averaged returns.}
}
\description{
General framework for testing for jumps on an intraday basis
}
\references{
COP2014, LM 2008
}
