% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump_tests.R
\name{intradayjumptest}
\alias{intradayjumptest}
\title{intradayjumptest}
\usage{
intradayjumptest(
  pdata,
  test.type = "LM",
  testing.times,
  window.size,
  K = 10,
  alpha = 0.05,
  extraArgs = list()
)
}
\description{
intradayjumptest
}
