% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump_tests.R
\name{intradayJumpTest}
\alias{intradayJumpTest}
\title{intradayJumpTest}
\usage{
intradayJumpTest(
  pData = NULL,
  testType = "LM",
  testingTimes,
  windowSize = 5,
  K = 10,
  alpha = 0.05,
  theta = 0.5,
  extraArgs = list(),
  ...
)
}
\description{
intradayJumpTest
}
