% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{knChooseReMeDI}
\alias{knChooseReMeDI}
\title{ReMeDI tuning parameter
function to choose the tuning parameter, kn in ReMeDI estimation}
\usage{
knChooseReMeDI(
  pData,
  correctTime = FALSE,
  jumpsIndex = NULL,
  knMax = 10,
  tol = 0.05,
  size = 3,
  lower = 2,
  upper = 5,
  plot = FALSE
)
}
\arguments{
\item{pData}{xts or data.table containing the log-prices of the asset.}

\item{correctTime}{logical indicating whether to use the time-adjusted ReMeDI measure, default is FALSE}

\item{jumpsIndex}{Indices of jump(s) detected}

\item{knMax}{max value of kn to be considered}

\item{tol}{tolerance for the minimizing value. If tol is high, the algorithm will choose a lower optimal value.}

\item{size}{size of the local window}

\item{lower}{lower boundary for the method if it fails to find an optimal value. If this is the case, the best kn between lower and upper is returned}

\item{upper}{upper boundary for the method if it fails to find an optimal value. If this is the case, the best kn between lower and upper is returned}

\item{plot}{logical whether to plot the errors.}
}
\value{
integer containing the optimal kn
}
\description{
ReMeDI tuning parameter
function to choose the tuning parameter, kn in ReMeDI estimation
}
\details{
This is the algorithm B.2 in the appendix of the Li and Linton (2019) working paper
}
\examples{
optimalKn <- knChooseReMeDI(sampleTDataMicroseconds, correctTime = FALSE, 
                            jumpsIndex = NULL, knMax = 10, tol = 0.05, 
                            size = 3, lower = 2, upper = 5, plot = TRUE)
optimalKn
\dontrun{
# We can also have a much larger search-space
optimalKn <- knChooseReMeDI(sampleTDataMicroseconds, correctTime = FALSE, 
                            jumpsIndex = NULL, knMax = 50, tol = 0.05,
                            size = 3, lower = 2, upper = 5, plot = TRUE)
optimalKn
}

}
\references{
A ReMeDI for Microstructure Noise
}
\author{
Emil Sjoerup
}
