% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadLag.R
\name{leadLag}
\alias{leadLag}
\title{Lead-Lag estimation}
\usage{
leadLag(
  price1 = NULL,
  price2 = NULL,
  lags = NULL,
  resolution = "seconds",
  normalize = TRUE,
  parallelize = FALSE,
  nCores = NA
)
}
\arguments{
\item{price1}{\code{xts} or \code{data.table} containing prices in levels, in case of data.table,
use a column DT to denote the date-time in POSIXct format, and a column PRICE to denote the price}

\item{price2}{\code{xts} or \code{data.table} containing prices in levels, in case of data.table,
use a column DT to denote the date-time in POSIXct format, and a column PRICE to denote the price}

\item{lags}{a numeric denoting which lags (in units of \code{resolution}) should be tested as leading or lagging}

\item{resolution}{the resolution at which the lags is measured. 
The default is "seconds", use this argument to gain 1000 times resolution by setting it to either "ms", "milliseconds", or "milli-seconds".}

\item{normalize}{logical denoting whether the contrasts should be normalized by the product of the L2 norms of both the prices. Default = TRUE. 
This does not change the value of the lead-lag-ratio.}

\item{parallelize}{logical denoting whether to use a parallelized version of the C++ code (parallelized using OPENMP). Default = FALSE}

\item{nCores}{integer valued numeric denoting how many cores to use for the lead-lag estimation procedure in case parallelize is TRUE. 
Default is NA, which does not parallelize the code.}
}
\value{
A list with class "leadLag" which containing 'contrasts', 'lead-lag-ratio', and 'lags', denoting the estimated values for each lag calculated,
the lead-lag-ratio, and the tested lags respectively.
}
\description{
Function that estimates whether one series leads (or lags) another.
}
\details{
The lead-lag-ratio (LLR) can be used to see if one asset leads the other. If LLR < 1, then price1 MAY be leading price2 and vice versa if LLR > 1.
}
\examples{
\dontrun{
# Toy example to show the usage
library(xts)
set.seed(123)
start <- strptime("1970-01-01", format = "\%Y-\%m-\%d", tz = "UTC")
timestamps <- start + seq(34200, 57600, length.out = 23401)

dat <- cbind(rnorm(23401) * sqrt(1/23401), rnorm(23401) * sqrt(1/23401))

dat <- exp(cumsum(xts(dat, timestamps)))
price1 <- dat[,1]
price2 <- dat[,2]
ll <- leadLag(price1, price2, seq(-10,10), normalize = FALSE)
plot(ll)
}

}
\references{
Hoffmann, M., Rosenbaum, M., and Yoshida, N. (2013). Estimation of the lead-lag parameter from non-synchronous data. \emph{Bernoulli}, 19, 1-37.
}
