% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadLag.R
\name{leadLag}
\alias{leadLag}
\title{Lead-Lag estimations
Function that}
\usage{
leadLag(
  price1,
  price2,
  lags,
  resolution = "seconds",
  normalize = TRUE,
  parallelize = FALSE,
  nCores = NA
)
}
\description{
Lead-Lag estimations
Function that
}
