% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{listAvailableKernels}
\alias{listAvailableKernels}
\title{Available Kernels}
\usage{
listAvailableKernels()
}
\value{
character vector
}
\description{
Returns a vector of the available kernels.
}
\examples{
listAvailableKernels
}
\references{
Barndorff-Nielsen, O. E. and Hansen, P. R. And Lunde, A. and Shephard, N. (2008). Designing Realized Kernels to Measure the ex post Variation of Equity Prices in the Presence of Noise. Econometrica, 76, pp. 1481-1536.
}
\author{
Scott Payseur
}
\keyword{volatility}
