% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realizedMeasures.R
\name{listAvailableKernels}
\alias{listAvailableKernels}
\title{Available Kernels}
\usage{
listAvailableKernels()
}
\value{
character vector
}
\description{
Returns a vector of the available kernels.
}
\examples{
listAvailableKernels
}
\references{
Barndorff-Nielsen, O. E., Hansen, P. R., Lunde, A., and Shephard, N. (2008). Designing realized kernels to measure the ex post variation of equity prices in the presence of noise. \emph{Econometrica}, 76, 1481-1536.
}
\author{
Scott Payseur
}
\keyword{volatility}
