% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{makeReturns}
\alias{makeReturns}
\title{Compute log returns}
\usage{
makeReturns(ts)
}
\arguments{
\item{ts}{xts object}
}
\value{
an xts object containing the log returns.
}
\description{
Function returns an xts object with the log returns as xts object.

\deqn{
\mbox{log return}_t =  (\log(\mbox{PRICE}_{t})-\log(\mbox{PRICE}_{t-1})).
}
}
\details{
Note: the first (row of) observation(s) is set to zero.
}
\author{
Jonathan Cornelissen and Kris Boudt
}
