% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{makeRMFormat}
\alias{makeRMFormat}
\title{Convert to format for realized measures}
\usage{
makeRMFormat(data)
}
\arguments{
\item{data}{An xts or a data.table object with atleast DT, SYMBOL, and PRICE columns. This data should already be cleaned.}
}
\value{
An xts or a data.table object with columns DT and a column named after each unique entrance in the SYMBOL column of the input. 
These columns contain the price of the associated symbol.
}
\description{
Convenience function to split data from one xts or data.table with atleast DT, SYMBOL, and PRICE columns to a format that can be used in the 
r* functions for calculation of realized measures.
}
\examples{
\dontrun{
library(data.table)
data1 <- copy(sampleTDataMicroseconds)[,  `:=`(PRICE = PRICE * runif(.N, min = 0.99, max = 1.01),
                                               DT = DT + runif(.N, 0.01, 0.02))]
data2 <- copy(sampleTDataMicroseconds)[, SYMBOL := 'XYZ']

dat <- rbind(data1, data2)
setkey(dat, "DT")
dat <- makeRMFormat(dat)

rCov(dat, alignBy = 'minutes', alignPeriod = 5, makeReturns = TRUE, cor = TRUE) 
}

}
