% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{medRQ}
\alias{medRQ}
\title{An estimator of integrated quarticity from applying the median operator on blocks of three returns.}
\usage{
medRQ(rData, alignBy = NULL, alignPeriod = NULL, makeReturns = FALSE)
}
\arguments{
\item{rData}{a zoo/xts object containing all returns in period t for one asset.}

\item{alignBy}{a string, align the tick data to "seconds"|"minutes"|"hours".}

\item{alignPeriod}{an integer, align the tick data to this many [seconds|minutes|hours].}

\item{makeReturns}{boolean, should be TRUE when rData contains prices instead of returns. FALSE by  default.}
}
\value{
numeric
}
\description{
Function returns the medRQ, defined in Andersen et al. (2012).
  
  Assume there is \eqn{N} equispaced returns in period \eqn{t}. Let \eqn{r_{t,i}} be a return (with \eqn{i=1, \ldots,N}) in period \eqn{t}.
  
  Then, the medRQ is given by
 \deqn{
   \mbox{medRQ}_{t}=\frac{3\pi N}{9\pi +72 - 52\sqrt{3}} \left(\frac{N}{N-2}\right) \sum_{i=2}^{N-1} \mbox{med}(|r_{t,i-1}|, |r_{t,i}|, |r_{t,i+1}|)^4
  }
}
\examples{
rq <- medRQ(rData = sampleTDataMicroseconds[, list(DT, PRICE)], alignBy = "minutes", 
            alignPeriod = 5, makeReturns = TRUE)
rq
}
\references{
Andersen, T. G., D. Dobrev, and E. Schaumburg (2012). Jump-robust volatility estimation using nearest neighbor truncation. Journal of Econometrics, 169(1), 75- 93.
}
\author{
Giang Nguyen, Jonathan Cornelissen and Kris Boudt
}
\keyword{highfrequency}
\keyword{medRQ}
