% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{noZeroPrices}
\alias{noZeroPrices}
\title{Delete the observations where the price is zero}
\usage{
noZeroPrices(tData)
}
\arguments{
\item{tData}{an \code{xts} or \code{data.table} object at least containing a column \code{PRICE}.}
}
\value{
an \code{xts} or \code{data.table} object depending on input
}
\description{
Function deletes the observations where the price is zero.
}
\author{
Jonathan Cornelissen and Kris Boudt
}
\keyword{cleaning}
