% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandling.R
\name{noZeroQuotes}
\alias{noZeroQuotes}
\title{Delete the observations where the bid or ask is zero}
\usage{
noZeroQuotes(qData)
}
\arguments{
\item{qData}{an \code{xts} or \code{data.table} object at least containing the columns \code{BID} and \code{OFR}.}
}
\value{
\code{xts} object or \code{data.table} depending on type of input.
}
\description{
Function deletes the observations where the bid or ask is zero.
}
\author{
Jonathan Cornelissen and Kris Boudt.
}
\keyword{cleaning}
