% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEAVYmodel.R
\name{plot.HEAVYmodel}
\alias{plot.HEAVYmodel}
\title{Plotting method for HEAVYmodel objects}
\usage{
\method{plot}{HEAVYmodel}(x, ...)
}
\arguments{
\item{x}{an object of class \code{HEAVYmodel}}

\item{...}{extra arguments, see details}
}
\description{
Plotting method for HEAVYmodel objects
}
\details{
The plotting method has the following optional parameter:
\itemize{
\item{\code{legend.loc}}{ A string denoting the location of the legend passed on to \code{addLegend} of the \pkg{xts} package}
}
}
