% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{quotesCleanup}
\alias{quotesCleanup}
\title{Cleans quote data}
\usage{
quotesCleanup(from, to, datasource, datadestination, ticker, exchanges,
  qdataraw = NULL, report = TRUE, selection = "median", maxi = 50,
  window = 50, type = "advanced", rmoutliersmaxi = 10)
}
\arguments{
\item{from}{character indicating first date to clean, e.g. "2008-01-30".}

\item{to}{character indicating last date to clean, e.g. "2008-01-31".}

\item{datasource}{character indicating the folder in which the original data is stored.}

\item{datadestination}{character indicating the folder in which the cleaned data is stored.}

\item{ticker}{vector of tickers for which the data should be cleaned, e.g. ticker = c("AAPL","AIG").}

\item{exchanges}{vector of stock exchange symbols for all tickers in vector "ticker". It thus should have the same length as the vector ticker.
Only data from one single exchange will be retained for each stock respectively, e.g. exchanges = c("T","N").
The possible exchange symbols are:
\itemize{
\item A: AMEX
\item N: NYSE
\item B: Boston
\item P: Arca
\item C: NSX
\item T/Q: NASDAQ
\item D: NASD ADF and TRF
\item X: Philadelphia
\item I: ISE
\item M: Chicago
\item W: CBOE
\item Z: BATS
}}

\item{qdataraw}{xts or data.table object containing (ONE stock only) raw quote data. This argument is NULL by default. Enabling it means the arguments
from, to, datasource and datadestination will be ignored. (only advisable for small chunks of data)}

\item{report}{boolean and TRUE by default. In case it is true the function returns (also) a vector indicating how many quotes remained after each cleaning step.}

\item{selection}{argument to be passed on to the cleaning routine \code{\link{mergeQuotesSameTimestamp}}. The default is "median".}

\item{maxi}{spreads which are greater than median(spreads of day) times maxi are excluded.}

\item{window}{argument to be passed on to the cleaning routine \code{\link{rmOutliersQuotes}}.}

\item{type}{argument to be passed on to the cleaning routine \code{\link{rmOutliersQuotes}}.}

\item{rmoutliersmaxi}{argument to be passed on to the cleaning routine \code{\link{rmOutliersQuotes}}.}
}
\value{
For each day an xts object is saved into the folder of that date, containing the cleaned data.
This procedure is performed for each stock in "ticker".
The function returns a vector indicating how many quotes remained after each cleaning step.

In case you supply the argument "qdataraw", the on-disk functionality is ignored
and the function returns a list with the cleaned quotes as an xts or data.table object depending on input (see examples).
}
\description{
This is a wrapper function for cleaning the quote data of all stocks in "ticker" over the interval [from,to]. 
The result is saved in the folder datadestination. 

In case you supply the argument "rawqdata", the on-disk functionality is ignored
and the function returns the cleaned quotes as xts or data.table object (see examples).

The following cleaning steps are performed sequentially:
\code{\link{noZeroQuotes}}, \code{\link{selectExchange}}, rmLargeSpread,
\code{\link{mergeQuotesSameTimestamp}}, \code{\link{rmOutliersQuotes}}.
}
\examples{
# Consider you have raw quote data for 1 stock for 2 days
head(sample_qdataraw_microseconds)
dim(sample_qdataraw_microseconds)
qdata_aftercleaning <- quotesCleanup(qdataraw = sample_qdataraw_microseconds, exchanges = "N")
qdata_aftercleaning$report
dim(qdata_aftercleaning$qdata)

# In case you have more data it is advised to use the on-disk functionality
# via "from","to","datasource",etc. arguments

}
\references{
Barndorff-Nielsen, O. E., P. R. Hansen, A. Lunde, and N. Shephard (2009). Realized kernels in practice: Trades and quotes. Econometrics Journal 12, C1-C32.
Brownlees, C.T. and Gallo, G.M. (2006). Financial econometric analysis at ultra-high frequency: Data handling concerns. Computational Statistics & Data Analysis, 51, pages 2232-2245.
Falkenberry, T.N. (2002). High frequency data filtering. Unpublished technical report.
}
\author{
Jonathan Cornelissen and Kris Boudt
}
\keyword{cleaning}
