% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{rAVGCov}
\alias{rAVGCov}
\title{Realized Covariance: Average Subsample}
\usage{
rAVGCov(
  rData,
  cor = FALSE,
  alignBy = "minutes",
  alignPeriod = 5,
  k = 1,
  makeReturns = FALSE
)
}
\arguments{
\item{rData}{an xts or data.table object containing returns or prices, possibly for multiple assets over multiple days
return series over period \eqn{t}, with \eqn{M} observations during \eqn{t}.}

\item{cor}{boolean, in case it is TRUE, the correlation is returned. FALSE by default.}

\item{alignBy}{Align the tick data to seconds|minutes|hours. Default is \code{"minutes"}}

\item{alignPeriod}{Align the tick data to this many [seconds|minutes|hours]. This can be a fraction. Default is \code{5}}

\item{k}{numeric denoting which horizon to use for the subsambles. This can be a fraction as long as k is a divisor of alignPeriod default is \code{1}}

\item{makeReturns}{Prices are passed make them into log returns}
}
\value{
in case the input is and contains data from one day, an N by N matrix is returned. If the data is a univariate xts object with multiple days, an xts is returned. 
If the data is multivariate and contains multiple days (xts or data.table), the function returns a list containing N by N matrices. Each item in the list has a name which corresponds to the date for the matrix.
}
\description{
Realized Covariance using average subsample.
}
\examples{
# Average subsampled realized variance/covariance aligned at one minute returns at 
# 5 subgrids (5 minutes).

# Univariate
rvSub <- rAVGCov(rData = sampleTData[, list(DT, PRICE)], alignBy = "minutes",
                 alignPeriod = 5, makeReturns = TRUE) 
rvSub

# Multivariate:
\dontrun{
library(xts)
set.seed(123)
start <- strptime("1970-01-01", format = "\%Y-\%m-\%d", tz = "UTC")
timestamps <- start + seq(34200, 57600, length.out = 23401)

dat <- cbind(rnorm(23401) * sqrt(1/23401), rnorm(23401) * sqrt(1/23401))

dat <- exp(cumsum(xts(dat, timestamps)))
price1 <- dat[,1]
price2 <- dat[,2]
rcovSub <- rAVGCov(rData = cbind(price1, price2), alignBy = "minutes", 
                   alignPeriod = 5, makeReturns = TRUE)
rcovSub

# Multivariate with a 30 second fast aggregation and a 2.5 minute slow aggregation.
rcovSub <- rAVGCov(rData = cbind(price1, price2), 
                   alignBy = "minutes", alignPeriod = 2.5, k = 0.5, makeReturns = TRUE)
rcovSub
}
}
\references{
L. Zhang, P.A Mykland, and Y. Ait-Sahalia. A tale of two time scales: Determining integrated volatility with noisy high-frequency data. \emph{Journal of the American Statistical Association}, 2005.

Michiel de Pooter, Martin Martens, and Dick van Dijk. Predicting the daily covariance matrix for S\&P100 stocks using intraday data - but which frequency to use? \emph{Econometric Reviews}, 2008.
}
\author{
Scott Payseur, Onno Kleen, and Emil Sjoerup
}
\keyword{volatility}
