% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{rCholCov}
\alias{rCholCov}
\title{Cholesky based realized covariance}
\usage{
rCholCov(
  pData,
  type = ".",
  criterion = "squared duration",
  delta = 0.1,
  theta = 1
)
}
\description{
Cholesky based realized covariance
}
