% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{rCov}
\alias{rCov}
\title{Realized Covariance}
\usage{
rCov(
  rData,
  cor = FALSE,
  alignBy = NULL,
  alignPeriod = NULL,
  makeReturns = FALSE
)
}
\arguments{
\item{rData}{an xts or data.table object containing returns or prices, possibly for multiple assets over multiple days}

\item{cor}{boolean, in case it is TRUE, the correlation is returned. FALSE by default.}

\item{alignBy}{character, indicating the time scale in which \code{alignPeriod} is expressed. Possible values are: "secs", "seconds", "mins", "minutes","hours".
To aggregate based on a 5 minute frequency, set \code{alignPeriod} to 5 and \code{alignBy} to \code{"minutes"}.}

\item{alignPeriod}{positive numeric, indicating the number of periods to aggregate over. E.g. to aggregate
based on a 5 minute frequency, set \code{alignPeriod} to 5 and \code{alignBy} to \code{"minutes"}.}

\item{makeReturns}{boolean, should be TRUE when rData contains prices instead of returns. FALSE by default.}
}
\value{
in case the input is and contains data from one day, an N by N matrix is returned. If the data is a univariate xts object with multiple days, an xts is returned. 
If the data is multivariate and contains multiple days (xts or data.table), the function returns a list containing N by N matrices. Each item in the list has a name which corresponds to the date for the matrix.
}
\description{
Function returns the Realized Covariation (rCov).
Let \eqn{r_{t,i}} be an intraday \eqn{N x 1} return vector and \eqn{i=1,...,M}
the number of intraday returns.

Then, the rCov is given by
\deqn{
 \mbox{rCov}_{t}=\sum_{i=1}^{M}r_{t,i}r'_{t,i}.
}
}
\examples{
# Realized Variance/Covariance for prices aligned at 5 minutes.

# Univariate: 
rv = rCov(rData = sampleTData[, list(DT, PRICE)], alignBy = "minutes", 
          alignPeriod = 5, makeReturns = TRUE)
rv

# Multivariate: 
rc = rCov(rData = sampleOneMinuteData, makeReturns=TRUE)
rc
}
\author{
Jonathan Cornelissen, Kris Boudt, and Emil Sjoerup
}
\keyword{volatility}
