% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{ReMeDI}
\alias{ReMeDI}
\title{ReMeDI
estimates auto-covariance of market-microstructure noise}
\usage{
ReMeDI(pData, kn = 1, lags = 1, knEqual = FALSE, makeCorrelation = FALSE)
}
\arguments{
\item{pData}{xts or data.table containing the log-prices of the asset}

\item{kn}{numeric of length 1 determining the tuning parameter kn this controls the lengths of the non-overlapping interval in the ReMeDI estimation}

\item{lags}{numeric containing integer values indicating}

\item{knEqual}{Use an altered version of the ReMeDI estimator, where we instead use equal kn, instead of kn and 2kn}

\item{makeCorrelation}{logical indicating whether to transform the autocovariances into autocorrelations}
}
\description{
ReMeDI
estimates auto-covariance of market-microstructure noise
}
\examples{
remed <- ReMeDI(sampleTDataMicroseconds[as.Date(DT) == "2018-01-02", ], kn = 5, lags = 1:8)
plot.ts(remed)
}
\references{
Li and Linton (2018) (Working paper): "A ReMeDI for microstructure noise."
}
\author{
Emil Sjoerup
}
\keyword{autocorrelation}
\keyword{autocovariance}
\keyword{microstructure}
\keyword{noise}
