% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{ReMeDI}
\alias{ReMeDI}
\title{ReMeDI
estimates auto-covariance of market-microstructure noise}
\usage{
ReMeDI(
  pData,
  kn = 1,
  lags = 1,
  correctTime = FALSE,
  jumpsIndex = NULL,
  makeCorrelation = FALSE
)
}
\arguments{
\item{pData}{xts or data.table containing the log-prices of the asset}

\item{kn}{numeric of length 1 determining the tuning parameter kn this controls the lengths of the non-overlapping interval in the ReMeDI estimation}

\item{lags}{numeric containing integer values indicating}

\item{correctTime}{logical indicating whether to use the time-adjusted ReMeDI measure, default is FALSE}

\item{jumpsIndex}{Indices of jump(s) detected}

\item{makeCorrelation}{logical indicating whether to transform the autocovariances into autocorrelations}
}
\description{
ReMeDI
estimates auto-covariance of market-microstructure noise
}
\references{
remedi paper, muzafer's paper
}
\author{
Emil Sjoerup
}
\keyword{autocorrelation}
\keyword{autocovariance}
\keyword{microstructure}
\keyword{noise}
