% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{rHYCov}
\alias{rHYCov}
\title{Hayashi-Yoshida Covariance}
\usage{
rHYCov(
  rData,
  cor = FALSE,
  period = 1,
  alignBy = "seconds",
  alignPeriod = 1,
  makeReturns = FALSE,
  makePsd = TRUE
)
}
\arguments{
\item{rData}{an xts or data.table object containing returns or prices, possibly for multiple assets.}

\item{cor}{boolean, in case it is TRUE, the correlation is returned. FALSE by default.}

\item{period}{Sampling period}

\item{alignBy}{Align the tick data to seconds|minutes|hours}

\item{alignPeriod}{Align the tick data to this many [seconds|minutes|hours]}

\item{makeReturns}{Prices are passed make them into log returns}

\item{makePsd}{boolean, in case it is TRUE, the positive definite version of rHYCov is returned. FALSE by default.}
}
\description{
Calculates the Hayashi-Yoshida Covariance estimator
}
\references{
T. Hayashi and N. Yoshida. On covariance estimation of non-synchronously observed diffusion processes. \emph{Bernoulli}, 11, 359-379, 2005.
}
\author{
Scott Payseur

# Average Hayashi-Yoshida Covariance estimator is calculated on five-minute returns

# Multivariate:
# realized_cov <- rHYCov(rData = cbind(lltc, sbux, fill = 0), period = 5, alignBy = "minutes", 
#                        alignPeriod = 5, makeReturns = FALSE)
# realized_cov 
Note: for the diagonal elements the rCov is used.
}
\keyword{volatility}
