% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{rHYCov}
\alias{rHYCov}
\title{Hayashi-Yoshida Covariance}
\usage{
rHYCov(
  rData,
  cor = FALSE,
  period = 1,
  alignBy = "seconds",
  alignPeriod = 1,
  makeReturns = FALSE,
  makePsd = TRUE
)
}
\arguments{
\item{rData}{an xts or data.table object containing returns or prices, possibly for multiple assets.}

\item{cor}{boolean, in case it is TRUE, the correlation is returned. FALSE by default.}

\item{period}{Sampling period}

\item{alignBy}{character, indicating the time scale in which \code{alignPeriod} is expressed. Possible values are: "secs", "seconds", "mins", "minutes","hours".
To aggregate based on a 5 minute frequency, set \code{alignPeriod} to 5 and \code{alignBy} to "minutes".}

\item{alignPeriod}{positive numeric, indicating the number of periods to aggregate over. E.g. to aggregate
based on a 5 minute frequency, set \code{alignPeriod} to 5 and \code{alignBy} to "minutes".}

\item{makeReturns}{Prices are passed make them into log returns}

\item{makePsd}{boolean, in case it is TRUE, the positive definite version of rHYCov is returned. FALSE by default.}
}
\description{
Calculates the Hayashi-Yoshida Covariance estimator
}
\references{
T. Hayashi and N. Yoshida. On covariance estimation of non-synchronously observed diffusion processes. \emph{Bernoulli}, 11, 359-379, 2005.
}
\author{
Scott Payseur and Emil Sjoerup

# Average Hayashi-Yoshida Covariance estimator is calculated on five-minute returns

# Multivariate:
# realized_cov <- rHYCov(rData = cbind(lltc, sbux, fill = 0), period = 5, alignBy = "minutes", 
#                        alignPeriod = 5, makeReturns = FALSE)
# realized_cov 
Note: for the diagonal elements the rCov is used.
}
\keyword{volatility}
