% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realizedMeasures.R
\name{rMedRQ}
\alias{rMedRQ}
\title{DEPRECATED}
\usage{
rMedRQ(rData, alignBy = NULL, alignPeriod = NULL, makeReturns = FALSE)
}
\arguments{
\item{rData}{DEPRECATED USE \code{\link{rMedRQuar}}}

\item{alignBy}{DEPRECATED USE \code{\link{rMedRQuar}}}

\item{alignPeriod}{DEPRECATED USE \code{\link{rMedRQuar}}}

\item{makeReturns}{DEPRECATED USE \code{\link{rMedRQuar}}}
}
\description{
DEPRECATED USE \code{\link{rMedRQuar}}
}
