% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realizedMeasures.R
\name{rMedRV}
\alias{rMedRV}
\title{DEPRECATED}
\usage{
rMedRV(rData, alignBy = NULL, alignPeriod = NULL, makeReturns = FALSE)
}
\arguments{
\item{rData}{DEPRECATED USE \code{\link{rMedRVar}}}

\item{alignBy}{DEPRECATED USE \code{\link{rMedRVar}}}

\item{alignPeriod}{DEPRECATED USE \code{\link{rMedRVar}}}

\item{makeReturns}{DEPRECATED USE \code{\link{rMedRVar}}}
}
\description{
DEPRECATED USE \code{\link{rMedRVar}}
}
