% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realizedMeasures.R
\name{rMinRV}
\alias{rMinRV}
\title{DEPRECATED}
\usage{
rMinRV(rData, alignBy = NULL, alignPeriod = NULL, makeReturns = FALSE)
}
\arguments{
\item{rData}{DEPRECATED USE \code{\link{rMinRVar}}}

\item{alignBy}{DEPRECATED USE \code{\link{rMinRVar}}}

\item{alignPeriod}{DEPRECATED USE \code{\link{rMinRVar}}}

\item{makeReturns}{DEPRECATED USE \code{\link{rMinRVar}}}
}
\description{
DEPRECATED USE \code{\link{rMinRVar}}
}
