% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{rmNegativeSpread}
\alias{rmNegativeSpread}
\title{Delete entries for which the spread is negative}
\usage{
rmNegativeSpread(qData)
}
\arguments{
\item{qData}{an xts object at least containing the columns "BID" and "OFR".}
}
\value{
data.table or xts object
}
\description{
Function deletes entries for which the spread is negative.
}
\examples{
rmNegativeSpread(sampleqDataRawMicroseconds)

}
\author{
Jonathan Cornelissen, Kris Boudt and Onno Kleen
}
\keyword{cleaning}
