% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{rQPVar}
\alias{rQPVar}
\title{Realized quad-power variation of highfrequency return series.}
\usage{
rQPVar(rData, alignBy = NULL, alignPeriod = NULL, makeReturns = FALSE)
}
\arguments{
\item{rData}{an xts or data.table object containing returns or prices, possibly for multiple assets over multiple days}

\item{alignBy}{character, indicating the time scale in which \code{alignPeriod} is expressed. Possible values are: "secs", "seconds", "mins", "minutes","hours".
To aggregate based on a 5 minute frequency, set \code{alignPeriod} to 5 and \code{alignBy} to "minutes".}

\item{alignPeriod}{positive numeric, indicating the number of periods to aggregate over. E.g. to aggregate
based on a 5 minute frequency, set \code{alignPeriod} to 5 and \code{alignBy} to "minutes".}

\item{makeReturns}{boolean, should be TRUE when rData contains prices instead of returns. FALSE by default.}
}
\value{
In case the input is an xts object with data from one day, a numeric of same length as the number of assets. 
If the input data spans multiple days and is in xts format, an xts will be returned. 
If the input data is a data.table object, the function returns a data.table with the same column names as the input data, containing the date and the realized measures
}
\description{
Function returns the realized quad-power variation, defined in Andersen et al. (2012).
 
 Assume there is \eqn{N} equispaced returns in period \eqn{t}. Let \eqn{r_{t,i}} be a return (with \eqn{i=1, \ldots,N}) in period \eqn{t}.
 
 Then, the rQPVar is given by
 \deqn{
   \mbox{rQPVar}_{t}=N*\frac{N}{N-3} \left(\frac{\pi^2}{4} \right)^{-4} \mbox({|r_{t,i}|} {|r_{t,i-1}|} {|r_{t,i-2}|} {|r_{t,i-3}|})
 }
}
\examples{
qpv <- rQPVar(rData= sampleTData[, list(DT, PRICE)], alignBy= "minutes", 
              alignPeriod =5, makeReturns= TRUE)
qpv

}
\references{
Andersen, T. G., D. Dobrev, and E. Schaumburg (2012). Jump-robust volatility estimation using nearest neighbor truncation. Journal of Econometrics, 169(1), 75- 93.
}
\author{
Giang Nguyen, Jonathan Cornelissen, Kris Boudt, and Emil Sjoerup
}
\keyword{highfrequency}
\keyword{rQPVar}
