% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realizedMeasures.R
\name{RV}
\alias{RV}
\title{DEPRECATED
DEPRECATED use \code{\link{rRVar}} instead}
\usage{
RV(rData)
}
\arguments{
\item{rData}{DEPRECATED \code{\link{rRVar}} instead}
}
\description{
DEPRECATED
DEPRECATED use \code{\link{rRVar}} instead
}
