% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_measures.R
\name{RV}
\alias{RV}
\title{An estimator of realized variance.}
\usage{
RV(rData)
}
\arguments{
\item{rData}{a \code{xts} object containing all returns in period t for one asset.}
}
\value{
numeric
}
\description{
An estimator of realized variance.
}
\keyword{RV}
\keyword{highfrequency}
