% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{salesCondition}
\alias{salesCondition}
\title{Delete entries with abnormal Sale Condition.}
\usage{
salesCondition(
  tData,
  validConds = c("", "@", "E", "@E", "F", "FI", "@F", "@FI", "I", "@I")
)
}
\arguments{
\item{tData}{an xts or data.table object containing the time series data, with 
one column named "COND" indicating the Sale Condition.}
}
\description{
Function deletes entries with abnormal Sale Condition: 
trades where column "COND" has
a letter code, except for "E" and "F".
}
\keyword{leaning}
