% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{salesCondition}
\alias{salesCondition}
\title{\link{salesCondition} is deprecated. Use \link{tradesCondition} instead.}
\usage{
salesCondition(
  tData,
  validConds = c("", "@", "E", "@E", "F", "FI", "@F", "@FI", "I", "@I")
)
}
\arguments{
\item{tData}{\link{salesCondition} is deprecated. Use \link{tradesCondition} instead.}

\item{validConds}{\link{salesCondition} is deprecated. Use \link{tradesCondition} instead.}
}
\description{
\link{salesCondition} is deprecated. Use \link{tradesCondition} instead.
}
\keyword{cleaning}
