% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_qdata}
\alias{sample_qdata}
\title{Sample of cleaned quotes for stock XXX for 1 day}
\format{xts object}
\usage{
sample_qdata
}
\description{
An xts object containing the raw quotes for the imaginary stock XXX for 1 day, in the typical NYSE TAQ database format. This is the cleaned version of the data sample \code{\link{sample_qdataraw}}, using \code{quotesCleanup}.
}
\keyword{datasets}
