% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleQDataraw}
\alias{sampleQDataraw}
\title{Sample of raw quotes for stock XXX for 1 day}
\format{
xts object
}
\usage{
sampleQDataraw
}
\description{
An imaginary xts object containing the raw quotes for stock XXX for 1 day, in the typical NYSE TAQ database format.
}
\keyword{datasets}
