\name{sample_qdataraw}
\docType{data}
\alias{sample_qdataraw}
\title{Sample of raw quotes for stock XXX for 1 day}
\description{
An imaginary xts object containing the raw quotes for stock XXX for 1 day, in the typical NYSE TAQ database format.
%This is the cleaned version of the data sample \code{\link{sample_tdataraw}}, using \code{\link{tradesCleanup}}.
}
\usage{data("sample_qdataraw")}
\format{xts object}
\keyword{datasets}