\name{sample_real5minprices}
\docType{data}
\alias{sample_real5minprices}
\title{Sample of imaginary price data for 61 days}
\description{
An xts object containing the 5-min aggregated imaginary price series for the trading
 days between 2005-03-04 and 2005-06-01.
 }
\usage{data("sample_real5minprices")}
\format{xts object}
\keyword{datasets}