% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_returns_5min}
\alias{sample_returns_5min}
\title{Sample returns data}
\format{A large \code{xts} object.}
\usage{
sample_returns_5min
}
\description{
EUR/USD returns from January to September 2004
}
\keyword{datasets}
