\name{sample_tdata}
\docType{data}
\alias{sample_tdata}
\title{Sample of cleaned trades for stock XXX for 1 day}
\description{
An xts object containing the trades for the imaginary stock XXX for 1 day, in the typical NYSE TAQ database format.
This is the cleaned version of the data sample \code{\link{sample_tdataraw}}, using \code{\link{tradesCleanup}}.
}
\usage{data("sample_tdata")}
\format{xts object}
\keyword{datasets}