% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampletDataMicroseconds}
\alias{sampletDataMicroseconds}
\title{Sample of cleaned trades for stock XXX for 2 days}
\format{
A data.table object.
}
\usage{
sampletDataMicroseconds
}
\description{
An data.table object containing the trades for the imaginary stock XXX for 2 days, in the typical NYSE TAQ database format.
This is the cleaned version of the data sample \code{\link{sampletDataRawMicroseconds}}, using \code{tradesCleanup}.
}
\keyword{datasets}
