% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_tdataraw}
\alias{sample_tdataraw}
\title{Sample of raw trades for stock XXX for 1 day}
\format{A large \code{xts} object.}
\usage{
sample_tdataraw
}
\description{
An imaginary xts object containing the raw trades for stock XXX for 1 day, in the typical NYSE TAQ database format.
This is the cleaned version of the data sample \code{\link{sample_tdataraw}}, using \code{tradesCleanup}.
}
\keyword{datasets}
