% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_tdataraw_microseconds}
\alias{sample_tdataraw_microseconds}
\title{Sample of raw trades for stock XXX for 2 days}
\format{A data.table object.}
\usage{
sample_tdataraw_microseconds
}
\description{
An imaginary data.table object containing the raw trades for stock XXX for 2 days, in the typical NYSE TAQ database format.
}
\keyword{datasets}
