% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleQData}
\alias{sampleQData}
\title{Sample of cleaned quotes for stock XXX for 2 days measured in microseconds}
\format{
data.table object
}
\usage{
sampleQData
}
\description{
A data.table object containing the quotes for the imaginary stock XXX for 2 days. This is the cleaned version of the data sample \code{\link{sampleQDataRaw}}, using \code{quotesCleanup}.
}
\examples{
\dontrun{
# The code to create the sampleQData dataset from raw data is
sampleQData <- quotesCleanup(qDataRaw = sampleQDataRaw,
                                         exchanges = "N", type = "standard", report = FALSE)
}
}
\keyword{datasets}
