% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_vol_and_drift.R
\name{spotDrift}
\alias{spotDrift}
\title{Spot Drift Estimation}
\usage{
spotDrift(
  data,
  method = "driftMean",
  alignBy = "minutes",
  alignPeriod = 5,
  marketOpen = "09:30:00",
  marketClose = "16:00:00",
  tz = NULL,
  ...
)
}
\arguments{
\item{data}{Can be one of two input types, \code{xts} or \code{data.table}. It is assumed that the input comprises prices in levels.}

\item{method}{Which method to be used to estimate the spot-drift. Currently, three methods are available, 
rolling mean and median as well as the kernel method of Christensen et al. 2018.
The kernel is a left hand exponential kernel that will weigh newer observations more heavily than older observations.}

\item{alignBy}{What time-frame should the estimator be applied? Accepted inputs are \code{"milliseconds"}, \code{"seconds"} and \code{"secs"} for seconds,
 \code{"minutes"} and \code{"mins"} for minutes, and \code{"hours"} for hours.
Standard is minutes}

\item{alignPeriod}{How often should the estimation take place? If \code{alignPeriod} is 5 the estimation will be done every fifth unit of \code{alignBy}.}

\item{marketOpen}{Opening time of the market, standard is "09:30:00"}

\item{marketClose}{Closing time of the market, standard is "16:00:00"}

\item{tz}{fallback time zone used in case we we are unable to identify the timezone of the data, by default: \code{tz = NULL}. We attempt to extract the timezone from the DT column (or index) of the data, which may fail. 
In case of failure we use \code{tz} if specified, and if it is not specified, we use \code{"UTC"}}

\item{...}{Additional arguments for the individual methods. See details}
}
\value{
An object of class \code{"spotDrift"} containing at least the estimated spot drift process. Input on what this class should contain and methods for it is welcome.
}
\description{
Function used to estimate the spot drift of intraday (tick) stock prices/returns
}
\details{
The additional arguments for the mean and median methods are: \code{periods} for the rolling window length which is 5 by standard and
\code{align} to allow for control of the alignment, should one wish to do so, the standard is \code{"right"}. 
For the kernel mean estimator, the arguments \code{meanBandwidth} can be used to control the bandwidth of the 
drift estimator and the \code{preAverage} argument, which can be used to control the pre-averaging horizon. 
These arguments default to 300 and 5 respectively.
}
\examples{
# Example 1: Rolling mean and median estimators for 2 days
meandrift <- spotDrift(data = sampleTData, alignPeriod = 1, tz = "EST")
mediandrift <- spotDrift(data = sampleTData, method = "driftMedian", 
                         alignBy = "seconds", alignPeriod = 30, tz = "EST")
plot(meandrift)
plot(mediandrift)
\dontrun{
# Example 2: Kernel based estimator for one day with data.table format
price <- sampleTData[as.Date(DT) == "2018-01-02", list(DT, PRICE)]
kerneldrift <- spotDrift(price, method = "driftKernel", alignBy = "minutes", alignPeriod = 1)
plot(kerneldrift)
}

}
\references{
Christensen, K., Oomen, R., and Reno, R. (2018) The Drift Burst Hypothesis (Working paper).
}
\author{
Emil Sjoerup
}
\keyword{Drift}
