% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_model.R
\name{summary.HARmodel}
\alias{summary.HARmodel}
\title{Summary for \code{HARmodel} objects}
\usage{
\method{summary}{HARmodel}(object)
}
\arguments{
\item{object}{An object of class \code{HARmodel}}
}
\value{
A modified \code{summary.lm}
}
\description{
Summary for \code{HARmodel} objects
}
